Index of functions and types

ODINN.FunctionalInversion — Type

mutable struct FunctionalInversion <: Simulation

An object representing a functional inversion simulation (i.e. the inversion of a function using some data-driven regressor).

Fields

model::Sleipnir.Model: The model used for the simulation.

glaciers::Vector{Sleipnir.AbstractGlacier}: A vector of glaciers involved in the simulation.

parameters::Sleipnir.Parameters: The parameters used for the simulation.

results::Vector{Results}: A vector to store the results of the simulation.

ODINN.FunctionalInversion — Method

function FunctionalInversion(
    model::Sleipnir.Model,
    glaciers::Vector{G},
    parameters::Sleipnir.Parameters
) where {G <: Sleipnir.AbstractGlacier}

Constructor for FunctionalInversion struct with glacier model information, glaciers, and parameters.

Arguments

model::Sleipnir.Model: The model used for the simulation.

glaciers::Vector{G}: A vector of glaciers involved in the simulation.

parameters::Sleipnir.Parameters: The parameters used for the simulation.

Returns

FunctionalInversion: A new instance of the FunctionalInversion struct.

ODINN.Hyperparameters — Method

Hyperparameters(;
    current_epoch::Int64 = nothing,
    current_minibatch::Int64 = nothing,
    loss_history::Vector{Float64} = Vector{Float64}(),
    optimizer::Optim.FirstOrderOptimizer = BFGS(initial_stepnorm=0.001),
    epochs::Int64 = 50,
    batch_size::Int64 = 15
    )

Initialize the hyperparameters of a machine learning model (Machine). Keyword arguments ================= - current_epoch: Current epoch in training - current_minibatch: Current minibatch in training - loss_history: Vector storing the loss for each epoch during training - optimizer: Optimizer to be used for training. Currently supports both Optim.jl and Flux.jl optimisers. - epochs: Number of epochs for the training - batch_size: Batch size for the training

ODINN.Inversion — Type

Inversion <: Simulation

A mutable struct that represents an inversion simulation.

Fields

model::Sleipnir.Model: The model used for the inversion.

glaciers::Vector{Sleipnir.AbstractGlacier}: A vector of glaciers involved in the inversion.

parameters::Sleipnir.Parameters: The parameters used for the inversion.

inversion::Vector{InversionResults}: A vector of results from the inversion.

ODINN.Inversion — Method

function Inversion(
    model::Sleipnir.Model,
    glaciers::Vector{Sleipnir.AbstractGlacier},
    parameters::Sleipnir.Parameters
    )

Construnctor for Inversion struct with glacier model infomation, glaciers and parameters. Keyword arguments =================

ODINN.InversionParameters — Type

InversionParameters{F<:AbstractFloat}

A mutable struct that holds parameters for inversion processes. This struct is a subtype of AbstractParameters.

Fields

initial_conditions::Vector{F}: A vector of initial conditions.

lower_bound::Vector{F}: A vector specifying the lower bounds for the parameters.

upper_bound::Vector{F}: A vector specifying the upper bounds for the parameters.

regions_split::Vector{Int}: A vector indicating how the regions are split.

x_tol::F: The tolerance for the solution's x-values.

f_tol::F: The tolerance for the function values.

solver::Any: The solver to be used for the inversion process.

ODINN.InversionParameters — Method

InversionParameters{F<:AbstractFloat}(;
    initial_conditions::Vector{F} = [1.0],
    lower_bound::Vector{F} = [0.0],
    upper_bound::Vector{F} = [Inf],
    regions_split::Vector{Int} = [1, 1],
    x_tol::F = 1.0e-3,
    f_tol::F = 1.0e-3,
    solver = BFGS()
)

Initialize the parameters for the inversion process.

Arguments

initial_conditions::Vector{F}: Starting point for optimization.

lower_bound::Vector{F}: Lower bounds for optimization variables.

upper_bound::Vector{F}: Upper bounds for optimization variables.

regions_split::Vector{Int}: Defines the amount of region split based on altitude and distance to border for the inversion process.

x_tol::F: Tolerance for variables convergence.

f_tol::F: Tolerance for function value convergence.

solver: Optimization solver to be used.

ODINN.InversionResults — Type

mutable struct InversionResults{F <: Real}

A structure to store the results of an inversion simulation.

Fields

rgi_id::Union{String, Nothing}: The RGI identifier, which can be a string or nothing.

A::F: Area parameter.

n::F: Exponent parameter.

C::Matrix{F}: Coefficient matrix.

H_pred::Matrix{F}: Predicted height matrix.

H_obs::Matrix{F}: Observed height matrix.

H_diff::Matrix{F}: Difference between predicted and observed height matrices.

V_pred::Matrix{F}: Predicted volume matrix.

V_obs::Matrix{F}: Observed volume matrix.

V_diff::Matrix{F}: Difference between predicted and observed volume matrices.

MSE::F: Mean squared error.

Δx::F: Grid spacing in the x-direction.

Δy::F: Grid spacing in the y-direction.

lon::F: Longitude.

lat::F: Latitude.

ODINN.NN — Type

NN{F <: AbstractFloat}(architecture::Flux.Chain, NN_f::Optimisers.Restructure, θ::Vector{F})

Feed-forward neural network.

Fields

architecture: Flux.Chain neural network architecture

NN_f: Neural network restructuring

θ: Neural network parameters

ODINN.NN — Method

NN(params::Parameters;
    architecture::Union{Flux.Chain, Nothing} = nothing,
    θ::Union{Vector{AbstractFloat}, Nothing} = nothing)

Creates a new feed-forward neural network.

Keyword arguments

architecture: Flux.Chain neural network architecture (optional)

θ: Neural network parameters (optional)

ODINN.UDEparameters — Method

UDEparameters(;
    sensealg::SciMLBase.AbstractAdjointSensitivityAlgorithm = GaussAdjoint(autojacvec=EnzymeVJP()),
    optimization_method::String = "AD+AD",
    loss_type::String = "V",
    scale_loss::Bool = true
    target::String = "D"
    )

Initialize the parameters for the training of the UDE. Keyword arguments ================= - sensealg: Sensitivity algorithm from SciMLSensitivity.jl to be used. - optimization_method: Optimization method for the UDE. - loss_type: Type of loss function to be used. Can be either V (ice velocities), or H (ice thickness). - scale_loss: Determines if the loss function should be scaled or not.

ODINN.Model — Method

function Model(; iceflow::Union{IFM, Nothing}, massbalance::Union{MBM, Nothing} machinelearning::Union{MLM, Nothing}, ) where {IFM <: IceflowModel, MBM <: MBmodel, MLM <: MLmodel}

Initialize Model at ODINN level (iceflow + mass balance + machine learning).

ODINN.Parameters — Method

Parameters(; physical::PhysicalParameters = PhysicalParameters(), simulation::SimulationParameters = SimulationParameters(), solver::SolverParameters = SolverParameters(), hyper::Hyperparameters = Hyperparameters(), UDE::UDEparameters = UDEparameters() inversion::InversionParameters = InversionParameters() ) Initialize ODINN parameters

Keyword arguments

physical::PhysicalParameters: Physical parameters for the simulation.

simulation::SimulationParameters: Parameters related to the simulation setup.

solver::SolverParameters: Parameters for the solver configuration.

hyper::Hyperparameters: Hyperparameters for the model.

UDE::UDEparameters: Parameters specific to the UDE (Universal Differential Equation).

inversion::InversionParameters: Parameters for inversion processes.

ODINN.enable_multiprocessing — Method

Initialize_ODINN(processes, python_path)

Initializes ODINN by configuring PyCall based on a given Python path. It also configures multiprocessing for a given number of processes.

ODINN.generate_batches — Method

generate_batches(simulation::S; shuffle=true)

Generates batches for the UE inversion problem based on input data and feed them to the loss function.

ODINN.get_NN — Method

get_NN(θ_trained)

Generates a neural network.

Arguments

θ_trained: Pre-trained neural network parameters (optional).

Returns

UA: Flux.Chain neural network architecture.

θ: Neural network parameters.

UA_f: Neural network restructuring.

ODINN.predict_A̅ — Method

predict_A̅(U, temp)

Predicts the value of A with a neural network based on the long-term air temperature.

ODINN.run! — Method

run!(simulation::FunctionalInversion)

In-place run of the model.

ODINN.sigmoid_A — Method

sigmoid_A(x)

Sigmoid activation function for the neural network output.

Arguments

x: Input value.

Returns

Sigmoid-transformed output value.

ODINN.sigmoid_A_inv — Method

sigmoid_A_inv(x)

Inverse sigmoid activation function for the neural network output.

Arguments

x: Input value.

Returns

Inverse sigmoid-transformed output value.

ODINN.simulate_iceflow_UDE! — Method

function simulateiceflowUDE!( θ, simulation::SIM, model::Sleipnir.Model, params::Sleipnir.Parameters, cb::DiscreteCallback, batch_id::I; du = Huginn.SIA2D) where {I <: Integer, SIM <: Simulation}

Make forward simulation of the iceflow UDE determined in du.

ODINN.train_UDE! — Method

train_UDE!(simulation::FunctionalInversion)

Trains UDE based on the current FunctionalInversion.

ODINN.update_training_state! — Method

update_training_state!(simulation::S, l) where {S <: Simulation}

Update the training state to determine if the training has completed an epoch. If an epoch is completed, reset the minibatches, update the history loss, and increment the epoch count.

Arguments

simulation: The current state of the simulation or training process.

l: The current loss value or other relevant metric.

Returns

None. This function updates the state in place.

Huginn.Prediction — Method

function Prediction(
    model::Sleipnir.Model,
    glaciers::Vector{Sleipnir.AbstractGlacier},
    parameters::Sleipnir.Parameters
    )

Construnctor for Prediction struct with glacier model infomation, glaciers and parameters. Keyword arguments =================

Huginn.SolverParameters — Method

SolverParameters(;
    solver::OrdinaryDiffEq.OrdinaryDiffEqAdaptiveAlgorithm = RDPK3Sp35(),
    reltol::Float64 = 1e-7
    )

Initialize the parameters for the numerical solver. Keyword arguments ================= - solver: solver to use from DifferentialEquations.jl - reltol: Relative tolerance for the solver

Huginn.SIA2D! — Method

SIA2D!(dH, H, SIA2Dmodel)

Compute an in-place step of the Shallow Ice Approximation PDE in a forward model

Huginn.SIA2D — Method

SIA(H, SIA2Dmodel)

Compute a step of the Shallow Ice Approximation UDE in a forward model. Allocates memory.

Huginn.avg — Method

avg(A)

4-point average of a matrix

Huginn.avg_surface_V! — Method

avg_surface_V(simulation::SIM)

Computes the average ice surface velocity for a given glacier evolution period based on the initial and final ice thickness states.

Huginn.avg_surface_V — Method

avg_surface_V(context, H, temp, sim, θ=[], UA_f=[])

Computes the average ice surface velocity for a given glacier evolution period based on the initial and final ice thickness states.

Huginn.avg_x — Method

avg_x(A)

2-point average of a matrix's X axis

Huginn.avg_y — Method

avg_y(A)

2-point average of a matrix's Y axis

Huginn.batch_iceflow_PDE! — Method

batch_iceflow_PDE!(glacier_idx::I, simulation::Prediction)

Solve the Shallow Ice Approximation iceflow PDE for a given temperature series batch in-place.

Huginn.batch_iceflow_PDE — Method

batch_iceflow_PDE(glacier_idx::I, simulation::Prediction)

Solve the Shallow Ice Approximation iceflow PDE for a given temperature series batch out-of-place.

Huginn.define_callback_steps — Method

define_callback_steps(tspan::Tuple{Float64, Float64}, step::Float64)

Defines the times to stop for the DiscreteCallback given a step

Huginn.diff_x — Method

diff_x(A)

2-point differential of a matrix's X axis

Huginn.diff_y — Method

diff_y(A)

2-point differential of a matrix's Y axis

Huginn.initialize_iceflow_model! — Method

function initializeiceflowmodel!(iceflowmodel::IF, glacieridx::I, glacier::AbstractGlacier, params::Sleipnir.Parameters ) where {IF <: IceflowModel, I <: Integer}

Initialize iceflow model data structures to enable in-place mutation.

Keyword arguments

- `iceflow_model`: Iceflow model used for simulation. 
- `glacier_idx`: Index of glacier.
- `glacier`: `Glacier` to provide basic initial state of the ice flow model.
- `parameters`: `Parameters` to configure some physical variables.

Huginn.initialize_iceflow_model — Method

function initializeiceflowmodel(iceflowmodel::IF, glacieridx::I, glacier::AbstractGlacier, params::Sleipnir.Parameters ) where {IF <: IceflowModel, I <: Integer}

Initialize iceflow model data structures to enable out-of-place mutation.

Keyword arguments

- `iceflow_model`: Iceflow model used for simulation. 
- `glacier_idx`: Index of glacier.
- `glacier`: `Glacier` to provide basic initial state of the ice flow model.
- `parameters`: `Parameters` to configure some physical variables.

Huginn.inn — Method

inn(A)

Access the inner part of the matrix (-2,-2)

Huginn.inn1 — Method

inn1(A)

Access the inner part of the matrix (-1,-1)

Huginn.run! — Method

run!(simulation::Prediction)

In-place run of the model.

Huginn.run₀ — Method

run(simulation::Prediction)

Out-of-place run of the model.

Huginn.simulate_iceflow_PDE! — Method

function simulate_iceflow_PDE!(
    simulation::SIM, 
    model::Sleipnir.Model, 
    params::Sleipnir.Parameters, 
    cb::DiscreteCallback; 
    du = SIA2D!) where {SIM <: Simulation}

Make forward simulation of the iceflow PDE determined in du.

Huginn.simulate_iceflow_PDE — Method

function simulate_iceflow_PDE(
    simulation::SIM, 
    model::Sleipnir.Model, 
    params::Sleipnir.Parameters, 
    cb::DiscreteCallback; 
    du = SIA2D) where {SIM <: Simulation}

Make forward simulation of the iceflow PDE determined in du.

Huginn.surface_V! — Method

surface_V!(H, B, Δx, Δy, temp, sim, A_noise, θ=[], UA_f=[])

Computes the ice surface velocity for a given glacier state

Huginn.surface_V — Method

surface_V(H, B, Δx, Δy, temp, sim, A_noise, θ=[], UA_f=[])

Computes the ice surface velocity for a given glacier state

Muninn.TImodel1 — Method

TImodel1(params::Sleipnir.Parameters;
    DDF::Float64 = 7.0/1000.0,
    acc_factor::Float64 = 1.0/1000.0
    )

Temperature-index model with a single degree-day factor.

Keyword arguments

- `DDF`: Single degree-day factor, for both snow and ice.
- `acc_factor`: Accumulation factor

Muninn.TImodel2 — Method

TImodel2(params::Parameters;
    DDF_snow::Float64 = 3.0/1000.0,
    DDF_ice::Float64 = 6.0/1000.0,
    acc_factor::Float64 = 1.0/1000.0
    )

Temperature-index model with two melt factors, for snow and ice.

Keyword arguments

- `DDF_snow`: Degree-day factor for snow.
- `DDF_ice`: Degree-day factor for ice.
- `acc_factor`: Accumulation factor

Muninn.MB_timestep! — Method

MB_timestep!(model::Model, glacier::G, step::F, t; batch_id::Union{Nothing, I} = nothing) where {I <: Integer, F <: AbstractFloat, G <: AbstractGlacier}

Retrieve the climate data, apply downscaling and compute the mass balance of a glacier for a specific time step and possibly for a batch of iceflow models.

Muninn.MB_timestep — Method

MB_timestep(model::Model, glacier::G, step::F, t::F) where {F <: AbstractFloat, G <: AbstractGlacier}

Retrieve the climate data, apply downscaling and compute the mass balance of a glacier for a specific time step.

Muninn.compute_MB — Method

compute_MB(mb_model::TImodel1, climate_2D_period::Climate2Dstep)

Compute the the mass balance given a mass balance model and a climate step.

Keyword arguments

- `mb_model`: Mass balance model.
- `climate_2D_period`: Climate step.

Sleipnir.Glacier1D — Method

function Glacier1D(; rgiid::Union{String, Nothing} = nothing, climate::Union{Climate1D, Nothing} = nothing, H₀::Union{Vector{F}, Nothing} = nothing, S::Union{Vector{F}, Nothing} = nothing, B::Union{Vector{F}, Nothing} = nothing, V::Union{Vector{F}, Nothing}= nothing, A::Union{F, Nothing} = nothing, C::Union{F, Nothing} = nothing, n::Union{F, Nothing} = nothing, w₀::Union{Vector{F}, Nothing} = nothing, λ::Union{Vector{F}, Nothing} = nothing, slope::Union{Vector{F}, Nothing} = nothing, distborder::Union{Vector{F}, Nothing} = nothing, S_coords::Union{Dict{String, Vector{Float64}}, Nothing} = nothing, Δx::Union{F, Nothing} = nothing, Δy::Union{F, Nothing} = nothing, nx::Union{I, Nothing} = nothing, ny::Union{I, Nothing} = nothing ) where {F <: AbstractFloat, I <: Integer}

Constructor for empty 2D Glacier object.

Sleipnir.Glacier2D — Method

function Glacier2D(; rgiid::Union{String, Nothing} = nothing, climate::Union{Climate2D, Nothing} = nothing, H₀::Union{Matrix{F}, Nothing} = nothing, Hglathida::Union{Matrix{F}, Nothing} = nothing, S::Union{Matrix{F}, Nothing} = nothing, B::Union{Matrix{F}, Nothing} = nothing, V::Union{Matrix{F}, Nothing}= nothing, Vx::Union{Matrix{F}, Nothing}= nothing, Vy::Union{Matrix{F}, Nothing}= nothing, A::Union{F, Nothing} = nothing, C::Union{F, Nothing} = nothing, n::Union{F, Nothing} = nothing, slope::Union{Matrix{F}, Nothing} = nothing, distborder::Union{Matrix{F}, Nothing} = nothing, Scoords::Union{Dict{String, Vector{Float64}}, Nothing} = nothing, Δx::Union{F, Nothing} = nothing, Δy::Union{F, Nothing} = nothing, nx::Union{I, Nothing} = nothing, ny::Union{I, Nothing} = nothing, cenlon::Union{F, Nothing} = nothing, cenlat::Union{F, Nothing} = nothing ) where {F <: AbstractFloat, I <: Integer}

Constructor for empty 2D Glacier object.

Sleipnir.Parameters — Method

Parameters(; physical::PhysicalParameters = PhysicalParameters(), simulation::SimulationParameters = SimulationParameters() ) Initialize ODINN parameters

Keyword arguments

Sleipnir.PhysicalParameters — Method

PhysicalParameters(;
    ρ::Float64 = 900.0,
    g::Float64 = 9.81,
    ϵ::Float64 = 1e-3,
    η₀::F = 1.0, 
    maxA::Float64 = 8e-17,
    minA::Float64 = 8.5e-20,
    maxTlaw::Float64 = 1.0,
    minTlaw::Float64 = -25.0,
    noise_A_magnitude::Float64 = 5e-18
    )

Initialize the physical parameters of a model. Keyword arguments ================= - ρ: Ice density - g: Gravitational constant - n: Glen's exponent - A: Glen's coefficient - ϵ: Small number - C: Sliding coefficient - η₀: - maxA: Maximum value for A (Glen's coefficient) - minA: Minimum value for A (Glen's coefficient)

Sleipnir.SimulationParameters — Method

SimulationParameters(;
                    use_MB::Bool = true,
                    use_iceflow::Bool = true,
                    plots::Bool = true,
                    velocities::Bool = true,
                    overwrite_climate::Bool = false,
                    use_glathida_data::Bool = false,
                    float_type::DataType = Float64,
                    int_type::DataType = Int64,
                    tspan::Tuple{F, F} = (2010.0,2015.0),
                    step::F = 1/12,
                    multiprocessing::Bool = true,
                    workers::I = 4,
                    working_dir::String = "",
                    test_mode::Bool = false,
                    rgi_paths::Dict{String, String} = Dict{String, String}(),
                    ice_thickness_source::String = "Farinotti19",
    )

Initialize the parameters for a simulation. Keyword arguments ================= - use_MB: Determines if surface mass balance should be used. - plots: Determines if plots should be made. - overwrite_climate: Determines if climate data should be overwritten - 'useglathidadata': Determines if data from the Glathida data set should be used

Sleipnir.apply_t_grad! — Method

apply_t_grad!(climate::RasterStack, dem::Raster)

Applies temperature gradients to the glacier 2D climate data based on a DEM.

Sleipnir.create_results — Method

store_results!(simulation::SIM, glacier_idx::I, solution) where {SIM <: Simulation, I <: Int}

Store the results of a simulation of a single glacier into a Results.

Sleipnir.downscale_2D_climate! — Method

downscale_2D_climate(glacier::Glacier2D)

Projects climate data to the glacier matrix by simply copying the closest gridpoint to all matrix gridpoints. Generates a new RasterStack which is returned.

Sleipnir.fillNaN! — Function

fillNaN!(x, fill)

Convert empty matrix grid cells into fill value

Sleipnir.get_cumulative_climate! — Function

get_cumulative_climate(climate, gradient_bounds=[-0.009, -0.003], default_grad=-0.0065)

Computes Positive Degree Days (PDDs) and cumulative rainfall and snowfall from climate data.

Sleipnir.get_raw_climate_data — Method

get_raw_climate_data(rgi_path::String)

Load the netCDF file containing the climate data for that glacier.

Sleipnir.initialize_glacier — Method

initialize_glacier(rgi_id::String, parameters::Parameters; smoothing=false, velocities=true)

Initialize a single Glaciers, including its Climate, based on a rgi_id and timestepping arguments.

Keyword arguments

- `rgi_id`: Glacier RGI ID
- `parameters`: Parameters including the physical and simulation ones
- `smoothing` Flag determining if smoothing needs to be applied to the surface elevation and ice thickness.
- `velocities` Flag determining if the ice surface velocities need to be retrieved.

Sleipnir.initialize_glacier_climate! — Method

function initialize_glacier_climate!(glacier::AbstractGlacier, params::Parameters)

Initializes the Climate data structure for a given Glacier`

Sleipnir.initialize_glacier_data — Method

initialize_glacier(rgi_id::String, params::Parameters; smoothing=false, velocities=true)

Retrieves the initial glacier geometry (bedrock + ice thickness) for a glacier with other necessary data (e.g. grid size and ice surface velocities).

Sleipnir.initialize_glaciers — Method

initialize_glaciers(rgi_ids::Vector{String}, params::Parameters; test=false)

Initialize multiple Glaciers based on a list of RGI IDs and on parameters.

Keyword arguments

- `rgi_ids`: List of RGI IDs of glaciers
- `params`: `Parameters` object to be passed

Sleipnir.plot_glacier — Method

plot_glacier(results::T, plot_type::String, variables::Vector{Symbol}; kwargs...) -> Figure

Generate various types of plots for glacier data.

Arguments

results: A custom type containing the results of a glacier simulation.

plot_type: Type of plot to generate. Options are:

"heatmaps": Heatmaps for glacier variables like :H, :S, :B, :V, :Vx, and :Vy.

"difference": Temporal difference metrics (between start and end) for a variable, with optional metrics like "hist" (histogram) and "difference".

"statistics": Temporal statistical metrics for a variable, with optional metrics like "average", "median", "min", "max", and "std".

"integrated_volume": Temporal evolution of the integrated ice volume for a variable.

variables: Variables to be plotted, e.g., :H.

Optional Keyword Arguments

tspan: A tuple representing the start and end time for the simulation.

metrics: Metrics to visualize, e.g., ["average"] for statistics, ["difference"] for difference.

Returns

A Figure object containing the desired visualization.

Notes

Ensure the variables and kwargs match the requirements of the specified plot_type.

The function routes requests to specific plotting functions based on plot_type.

Sleipnir.save_results_file! — Method

save_results_file(simulation::Prediction)

Save simulation Results into a .jld2 file.

Sleipnir.smooth! — Method

smooth!(A)

Smooth data contained in a matrix with one time step (CFL) of diffusion.

Sleipnir.stop_condition_tstops — Method

stop_condition_tstops(u,t,integrator, tstops)

Function that iterates through the tstops, with a closure including tstops

Sleipnir.trim_period — Method

trim_period(period, climate)

Trims a time period based on the time range of a climate series.