Index of functions and types

Parameters

There are different types of parameters, holding specific information for different modelling aspects. All the types of parameters are wrapped into a Parameter type, which is threaded throughout ODINN.jl.

Sleipnir.SimulationParameters — Type

SimulationParameters(;
                    use_MB::Bool = true,
                    use_iceflow::Bool = true,
                    plots::Bool = true,
                    velocities::Bool = true,
                    overwrite_climate::Bool = false,
                    use_glathida_data::Bool = false,
                    tspan::Tuple{F, F} = (2010.0,2015.0),
                    step::F = 1/12,
                    multiprocessing::Bool = true,
                    workers::I = 4,
                    working_dir::String = "",
                    test_mode::Bool = false,
                    rgi_paths::Dict{String, String} = Dict{String, String}(),
                    ice_thickness_source::String = "Farinotti19",
    )

Initialize the parameters for a simulation. Keyword arguments ================= - use_MB: Determines if surface mass balance should be used. - plots: Determines if plots should be made. - overwrite_climate: Determines if climate data should be overwritten - 'useglathidadata': Determines if data from the Glathida data set should be used

Sleipnir.PhysicalParameters — Type

PhysicalParameters(;
        ρ::F = 900.0,
        g::F = 9.81,
        ϵ::F = 1e-3,
        η₀::F = 1.0, 
        maxA::F = 8e-17,
        minA::F = 8.5e-20,
        maxTlaw::F = 1.0,
        minTlaw::F = -25.0,
        noise_A_magnitude::F = 5e-18
        )

Initialize the physical parameters of a model. Keyword arguments ================= - ρ: Ice density - g: Gravitational constant - ϵ: Small number - η₀: - maxA: Maximum value for A (Glen's coefficient) - minA: Minimum value for A (Glen's coefficient) - maxTlaw: Maximum value of Temperature used in simulations on fake law - minTlaw: Minimum value of Temperature used in simulations on fake law - noise_A_magnitude: Magnitude of noise added to A

ODINN.Hyperparameters — Type

Hyperparameters(;
    current_epoch::Int64 = nothing,
    current_minibatch::Int64 = nothing,
    loss_history::Vector{Float64} = Vector{Float64}(),
    optimizer::Optim.FirstOrderOptimizer = BFGS(initial_stepnorm=0.001),
    epochs::Int64 = 50,
    batch_size::Int64 = 15
    )

Initialize the hyperparameters of a machine learning model (Machine). Keyword arguments ================= - current_epoch: Current epoch in training - current_minibatch: Current minibatch in training - loss_history: Vector storing the loss for each epoch during training - optimizer: Optimizer to be used for training. Currently supports both Optim.jl and Flux.jl optimisers. - epochs: Number of epochs for the training - batch_size: Batch size for the training

ODINN.UDEparameters — Type

UDEparameters(;
    sensealg::SciMLBase.AbstractAdjointSensitivityAlgorithm = GaussAdjoint(autojacvec=EnzymeVJP()),
    optimization_method::String = "AD+AD",
    loss_type::String = "V",
    scale_loss::Bool = true
    target::Union{String, Nothing} = "D"
    )

Initialize the parameters for the training of the UDE. Keyword arguments ================= - sensealg: Sensitivity algorithm from SciMLSensitivity.jl to be used. - optimization_method: Optimization method for the UDE. - loss_type: Type of loss function to be used. Can be either V (ice velocities), or H (ice thickness). - scale_loss: Determines if the loss function should be scaled or not.

All these subtypes of parameters are gathered in a Parameters struct:

ODINN.Parameters — Function

Parameters(; physical::PhysicalParameters = PhysicalParameters(), simulation::SimulationParameters = SimulationParameters(), solver::SolverParameters = SolverParameters(), hyper::Hyperparameters = Hyperparameters(), UDE::UDEparameters = UDEparameters() inversion::InversionParameters = InversionParameters() ) Initialize ODINN parameters

Keyword arguments

physical::PhysicalParameters: Physical parameters for the simulation.

simulation::SimulationParameters: Parameters related to the simulation setup.

solver::SolverParameters: Parameters for the solver configuration.

hyper::Hyperparameters: Hyperparameters for the model.

UDE::UDEparameters: Parameters specific to the UDE (Universal Differential Equation).

inversion::InversionParameters: Parameters for inversion processes.

Glaciers

Glaciers in ODINN.jl are represented by a Glacier type. Each glacier has its related Climate. Since ODINN.jl supports different types of simulations, we offer the possibility to work on 1D (i.e. flowline), 2D (e.g. SIA) or even 3D (not yet implemented, e.g. Full Stokes).

Sleipnir.Glacier2D — Type

function Glacier2D(; rgiid::Union{String, Nothing} = nothing, climate::Union{Climate2D, Nothing} = nothing, H₀::Union{Matrix{F}, Nothing} = nothing, Hglathida::Union{Matrix{F}, Nothing} = nothing, S::Union{Matrix{F}, Nothing} = nothing, B::Union{Matrix{F}, Nothing} = nothing, V::Union{Matrix{F}, Nothing}= nothing, Vx::Union{Matrix{F}, Nothing}= nothing, Vy::Union{Matrix{F}, Nothing}= nothing, A::Union{F, Nothing} = nothing, C::Union{F, Nothing} = nothing, n::Union{F, Nothing} = nothing, slope::Union{Matrix{F}, Nothing} = nothing, dist_border::Union{Matrix{F}, Nothing} = nothing, Coords::Union{Dict{String, Vector{Float64}}, Nothing} = nothing, Δx::Union{F, Nothing} = nothing, Δy::Union{F, Nothing} = nothing, nx::Union{I, Nothing} = nothing, ny::Union{I, Nothing} = nothing, cenlon::Union{F, Nothing} = nothing, cenlat::Union{F, Nothing} = nothing ) where {F <: AbstractFloat, I <: Integer}

Constructor for empty 2D Glacier object.

Every glacier has its associated climate, following the same spatial representation (e.g. 2D):

<!– @docs Sleipnir.Climate2D –>

Models

There are 3 main types of models in ODINN.jl, iceflow models, mass balance models and machine learning models.

Work in progress...