Index of functions and types

ODINN.AbstractAdjointMethod — Type

AbstractAdjointMethod

Abstract type representing the flavor of AD and adjoint to be used to compute the gradient of the cost function. There are two parts where one can play with how the gradient is propagated: the iceflow model VJP and the adjoint of the ODE solver. The VJP of the iceflow model can be computed using either AD (Zygote or Enzyme), the discrete, or the continuous adjoint of the iceflow model. As for the computation of the adjoint of the ODE solution, it can be handled by SciMLSensitivity, or computed using the adjoint implemented in ODINN.

ODINN.ContinuousAdjoint — Type

Continuous adjoint of SIA2D with manual implementation of the backward in the ODE scheme.

ContinuousAdjoint{F <: AbstractFloat}

Fields

reltol::F: Relative tolerance to be used in the ODE solver of the adjoint.

abstol::F: Absolute tolerance to be used in the ODE solver of the adjoint.

ODINN.DiscreteAdjoint — Type

Discrete adjoint of SIA2D with manual implementation of the backward in the ODE scheme.

DiscreteAdjoint

ODINN.DummyAdjoint — Type

Struct to provide a dummy gradient. It does not have to be the true gradient. Mainly used to test that the optimization pileline works independenly of the gradient calculation.

DummyAdjoint

Fields:

grad::Function: In-place function f(du, u; kwargs) that fills the first argument du with the gradient values.

ODINN.EnzymeAdjoint — Type

Enzyme AD of SIA2D with manual implementation of the backward in the ODE scheme.

EnzymeAdjoint

ODINN.FunctionalInversion — Type

mutable struct FunctionalInversion <: Simulation

An object representing a functional inversion simulation (i.e. the inversion of a function using some data-driven regressor).

Fields

model::Sleipnir.Model: The model used for the simulation.

glaciers::Vector{Sleipnir.AbstractGlacier}: A vector of glaciers involved in the simulation.

parameters::Sleipnir.Parameters: The parameters used for the simulation.

results::Vector{Results}: A vector to store the results of the simulation.

ODINN.FunctionalInversion — Method

function FunctionalInversion(
    model::Sleipnir.Model,
    glaciers::Vector{G},
    parameters::Sleipnir.Parameters
) where {G <: Sleipnir.AbstractGlacier}

Constructor for FunctionalInversion struct with glacier model information, glaciers, and parameters.

Arguments

model::Sleipnir.Model: The model used for the simulation.

glaciers::Vector{G}: A vector of glaciers involved in the simulation.

parameters::Sleipnir.Parameters: The parameters used for the simulation.

Returns

FunctionalInversion: A new instance of the FunctionalInversion struct.

ODINN.Hyperparameters — Method

Hyperparameters(;
    current_epoch::Int64 = nothing,
    current_minibatch::Int64 = nothing,
    loss_history::Vector{Float64} = Vector{Float64}(),
    optimizer::Optim.FirstOrderOptimizer = BFGS(initial_stepnorm=0.001),
    epochs::Int64 = 50,
    batch_size::Int64 = 15
    )

Initialize the hyperparameters of a machine learning model (Machine). Keyword arguments ================= - current_epoch: Current epoch in training - current_minibatch: Current minibatch in training - loss_history: Vector storing the loss for each epoch during training - optimizer: Optimizer to be used for training. Currently supports both Optim.jl and Flux.jl optimisers. - epochs: Number of epochs for the training - batch_size: Batch size for the training

ODINN.Inversion — Type

Inversion <: Simulation

A mutable struct that represents an inversion simulation.

Fields

model::Sleipnir.Model: The model used for the inversion.

glaciers::Vector{Sleipnir.AbstractGlacier}: A vector of glaciers involved in the inversion.

parameters::Sleipnir.Parameters: The parameters used for the inversion.

inversion::Vector{InversionResults}: A vector of results from the inversion.

ODINN.Inversion — Method

function Inversion(
    model::Sleipnir.Model,
    glaciers::Vector{Sleipnir.AbstractGlacier},
    parameters::Sleipnir.Parameters
    )

Construnctor for Inversion struct with glacier model infomation, glaciers and parameters. Keyword arguments =================

ODINN.InversionParameters — Type

InversionParameters{F<:AbstractFloat}

A mutable struct that holds parameters for inversion processes. This struct is a subtype of AbstractParameters.

Fields

initial_conditions::Vector{F}: A vector of initial conditions.

lower_bound::Vector{F}: A vector specifying the lower bounds for the parameters.

upper_bound::Vector{F}: A vector specifying the upper bounds for the parameters.

regions_split::Vector{Int}: A vector indicating how the regions are split.

x_tol::F: The tolerance for the solution's x-values.

f_tol::F: The tolerance for the function values.

solver::Any: The solver to be used for the inversion process.

ODINN.InversionParameters — Method

InversionParameters{F<:AbstractFloat}(;
    initial_conditions::Vector{F} = [1.0],
    lower_bound::Vector{F} = [0.0],
    upper_bound::Vector{F} = [Inf],
    regions_split::Vector{Int} = [1, 1],
    x_tol::F = 1.0e-3,
    f_tol::F = 1.0e-3,
    solver = BFGS()
)

Initialize the parameters for the inversion process.

Arguments

initial_conditions::Vector{F}: Starting point for optimization.

lower_bound::Vector{F}: Lower bounds for optimization variables.

upper_bound::Vector{F}: Upper bounds for optimization variables.

regions_split::Vector{Int}: Defines the amount of region split based on altitude and distance to border for the inversion process.

x_tol::F: Tolerance for variables convergence.

f_tol::F: Tolerance for function value convergence.

solver: Optimization solver to be used.

ODINN.InversionResults — Type

mutable struct InversionResults{F <: Real}

A structure to store the results of an inversion simulation.

Fields

rgi_id::Union{String, Nothing}: The RGI identifier, which can be a string or nothing.

A::F: Area parameter.

n::F: Exponent parameter.

C::Matrix{F}: Coefficient matrix.

H_pred::Matrix{F}: Predicted height matrix.

H_obs::Matrix{F}: Observed height matrix.

H_diff::Matrix{F}: Difference between predicted and observed height matrices.

V_pred::Matrix{F}: Predicted volume matrix.

V_obs::Matrix{F}: Observed volume matrix.

V_diff::Matrix{F}: Difference between predicted and observed volume matrices.

MSE::F: Mean squared error.

Δx::F: Grid spacing in the x-direction.

Δy::F: Grid spacing in the y-direction.

lon::F: Longitude.

lat::F: Latitude.

ODINN.NeuralNetwork — Type

NeuralNetwork{
    ChainType <: Lux.Chain,
    ComponentVectorType <: ComponentVector,
    NamedTupleType <: NamedTuple
} <: MLmodel

Feed-forward neural network.

Fields

architecture::ChainType: Flux.Chain neural network architecture

θ::ComponentVectorType: Neural network parameters

st::NamedTupleType: Neural network status

ODINN.NeuralNetwork — Method

NN(params::Parameters;
    architecture::Union{Flux.Chain, Nothing} = nothing,
    θ::Union{Vector{AbstractFloat}, Nothing} = nothing)

Creates a new feed-forward neural network.

Keyword arguments

architecture: Flux.Chain neural network architecture (optional)

θ: Neural network parameters (optional)

ODINN.SciMLSensitivityAdjoint — Type

Struct that defines the SciMLSensitivity adjoint flavor. This is the default behavior in ODINN.

SciMLSensitivityAdjoint

ODINN.TrainingStats — Type

mutable struct TrainingStats

An object with the information of the training.

Fields

retcode: Report code of the optimization.

losses: Vector storing the value of the loss function at each iteration.

niter: Total number of iterations/epochs.

θ: Parameters of neural network after training

ODINN.TrainingStats — Method

function TrainingStats(;
    retcode::Union{String, Nothing} = nothing,
    losses::Vector{F} = Float64[],
    niter::I = 0
) where {F <: AbstractFloat, I <: Int}

Constructor for TrainingStats object used to store important information during training.

Arguments

retcode: Report code of the optimization.

losses: Vector storing the value of the loss function at each iteration.

niter: Total number of iterations/epochs.

θ: Parameters of neural network after training

ODINN.UDEparameters — Method

UDEparameters(;
    sensealg::SciMLBase.AbstractAdjointSensitivityAlgorithm = GaussAdjoint(autojacvec=EnzymeVJP()),
    optimization_method::String = "AD+AD",
    loss_type::String = "V",
    scale_loss::Bool = true
    target::Union{String, Nothing} = "D"
    )

Initialize the parameters for the training of the UDE. Keyword arguments ================= - sensealg: Sensitivity algorithm from SciMLSensitivity.jl to be used. - optimization_method: Optimization method for the UDE. - loss_type: Type of loss function to be used. Can be either V (ice velocities), or H (ice thickness). - scale_loss: Determines if the loss function should be scaled or not.

ODINN.ZygoteAdjoint — Type

Zygote AD of SIA2D with manual implementation of the backward in the ODE scheme. This flavor is not supported yet.

ZygoteAdjoint

ODINN.A_law_PatersonCuffey — Method

A_law_PatersonCuffey()

Returns a law of the coefficient A as a polynomial of the temperature. The values used to fit the polynomial come from Peterson & Cuffey.

ODINN.CallbackOptimizationSet — Method

CallbackOptimizationSet(θ, l; callbacks)

Helper to combine callbacks for Optimization function. This executes the action of each callback. (equivalent to CallbackSet for DifferentialEquations.jl)

ODINN.Model — Method

function Model(; iceflow::Union{IFM, Nothing}, massbalance::Union{MBM, Nothing} machinelearning::Union{MLM, Nothing}, ) where {IFM <: IceflowModel, MBM <: MBmodel, MLM <: MLmodel}

Initialize Model at ODINN level (iceflow + mass balance + machine learning).

ODINN.Parameters — Method

Parameters(; physical::PhysicalParameters = PhysicalParameters(), simulation::SimulationParameters = SimulationParameters(), solver::SolverParameters = SolverParameters(), hyper::Hyperparameters = Hyperparameters(), UDE::UDEparameters = UDEparameters() inversion::InversionParameters = InversionParameters() ) Initialize ODINN parameters

Keyword arguments

physical::PhysicalParameters: Physical parameters for the simulation.

simulation::SimulationParameters: Parameters related to the simulation setup.

solver::SolverParameters: Parameters for the solver configuration.

hyper::Hyperparameters: Hyperparameters for the model.

UDE::UDEparameters: Parameters specific to the UDE (Universal Differential Equation).

inversion::InversionParameters: Parameters for inversion processes.

ODINN.SIA2D_adjoint! — Method

Define SIA2D forward map for the adjoint mode

ODINN.SIA2D_grad! — Method

Inverse with batch

ODINN.SIA2D_grad_batch! — Method

Inverse by glacier

ODINN.VJP_λ_∂SIA∂H_continuous — Method

VJP_λ_∂SIA∂H_continuous(
    λ::Matrix{R},
    H::Matrix{R},
    simulation::SIM,
    t::R;
    batch_id::Union{Nothing, I} = nothing
)

Implementation of the continuous adjoint of the SIA2D equation with respect to H. Given λ and H, it returns the VJP of λ^T * ∂(SIA2D)/∂H (H).

Arguments:

λ::Matrix{R}: Adjoint state, also called output gradient in reverse-mode AD.

H::Matrix{R}: Ice thickness which corresponds to the input state of the SIA2D.

simulation::SIM: Simulation parameters.

t::R: Time value, not used as SIA2D is time independent.

batch_id::Union{Nothing, I}: Batch index.

Returns:

dλ::Matrix{R}: Jacobian vector product, also called input gradient in reverse-mode AD.

ODINN.VJP_λ_∂SIA∂θ_continuous — Method

VJP_λ_∂SIA∂θ_continuous(
    θ,
    λ::Matrix{R},
    H::Matrix{R},
    simulation::SIM,
    t::R;
    batch_id::Union{Nothing, I} = nothing
)

Implementation of the continuous adjoint of the SIA2D equation with respect to θ. Given λ, H and θ, it returns the VJP of λ^T * ∂(SIA2D)/∂θ (θ).

Arguments:

θ: Vector of parameters

λ::Matrix{R}: Adjoint state, also called output gradient in reverse-mode AD.

H::Matrix{R}: Ice thickness which corresponds to the input state of the SIA2D.

simulation::SIM: Simulation parameters.

t::R: Time value, not used as SIA2D is time independent.

batch_id::Union{Nothing, I}: Batch index.

Returns:

dλ::Matrix{R}: Jacobian vector product, also called input gradient in reverse-mode AD.

ODINN.callback_diagnosis — Method

callback_diagnosis(θ, l, simulation)

Callback function to track and diagose training. It includes print and updates in simulation::Simulation.

ODINN.callback_plots_A — Method

callback_plots_A(θ, l, simulation)

Callback function to generate plots during training.

ODINN.generate_ground_truth — Method

generate_ground_truth(glacier::G, fakeA::Function, params, model, tstops::Vector{F})

Generates ground truth data and populate glacier with the ground truth observation given a fake law of A.

Arguments:

glacier::G: Glacier instance.

fakeA::Function: Function that maps a temperature to A.

params::: The simulation parameters.

model:: The model that includes iceflow and a machine learning model.

tstops:: Vector of time points where the solver should stop.

ODINN.generate_simulation_batches — Method

generate_batches(simulation::S; shuffle=true)

Generates batches for the UE inversion problem based on input data and feed them to the loss function.

ODINN.get_NN — Method

get_NN(θ_trained, ft)

Generates a neural network.

Arguments

ft: Float type.

θ_trained: Pre-trained neural network parameters (optional).

ft: Float type used.

Returns

UA: Lux.Chain neural network architecture.

θ: Neural network parameters.

st: Lux state.

ODINN.predict_A̅ — Method

predict_A̅(U, temp)

Predicts the value of A with a neural network based on the long-term air temperature.

ODINN.run! — Method

run!(simulation::FunctionalInversion)

In-place run of the model.

ODINN.sigmoid_A — Method

sigmoid_A(x)

Sigmoid activation function for the neural network output.

Arguments

x: Input value.

Returns

Sigmoid-transformed output value.

ODINN.sigmoid_A_inv — Method

sigmoid_A_inv(x)

Inverse sigmoid activation function for the neural network output.

Arguments

x: Input value.

Returns

Inverse sigmoid-transformed output value.

ODINN.simulate_iceflow_UDE! — Method

function simulateiceflowUDE!( θ, simulation::SIM, cb::DiscreteCallback, batch_id::I; du = Huginn.SIA2D) where {I <: Integer, SIM <: Simulation}

Make forward simulation of the iceflow UDE determined in du.

ODINN.train_UDE! — Method

BFGS Training

ODINN.train_UDE! — Method

train_UDE!(simulation::FunctionalInversion)

Trains UDE based on the current FunctionalInversion.

ODINN.train_UDE! — Method

ADAM Training

ODINN.update_training_state! — Method

update_training_state!(simulation::S, l) where {S <: Simulation}

Update the training state to determine if the training has completed an epoch. If an epoch is completed, reset the minibatches, update the history loss, and increment the epoch count.

Arguments

simulation: The current state of the simulation or training process.

l: The current loss value or other relevant metric.

Returns

None. This function updates the state in-place.

Huginn.IceflowModel — Type

IceflowModel

An abstract type representing the base model for ice flow simulations. All specific ice flow models should subtype this abstract type.

Huginn.Prediction — Method

function Prediction(
    model::Sleipnir.Model,
    glaciers::Vector{Sleipnir.AbstractGlacier},
    parameters::Sleipnir.Parameters
    )

Construnctor for Prediction struct with glacier model infomation, glaciers and parameters. Keyword arguments =================

Huginn.SIA2Dmodel — Type

mutable struct SIA2Dmodel{R <: Real, I <: Integer} <: SIAmodel

A mutable struct representing a 2D Shallow Ice Approximation (SIA) model.

Fields

A::Union{Ref{R}, Nothing}: Flow rate factor.

n::Union{Ref{R}, Nothing}: Flow law exponent.

C::Union{Ref{R}, Matrix{R}, Nothing}: Sliding coefficient.

H₀::Union{Matrix{R}, Nothing}: Initial ice thickness.

H::Union{Matrix{R}, Nothing}: Ice thickness.

H̄::Union{Matrix{R}, Nothing}: Averaged ice thickness.

S::Union{Matrix{R}, Nothing}: Surface elevation.

dSdx::Union{Matrix{R}, Nothing}: Surface slope in the x-direction.

dSdy::Union{Matrix{R}, Nothing}: Surface slope in the y-direction.

D::Union{Matrix{R}, Nothing}: Diffusivity.

Dx::Union{Matrix{R}, Nothing}: Diffusivity in the x-direction.

Dy::Union{Matrix{R}, Nothing}: Diffusivity in the y-direction.

dSdx_edges::Union{Matrix{R}, Nothing}: Surface slope at edges in the x-direction.

dSdy_edges::Union{Matrix{R}, Nothing}: Surface slope at edges in the y-direction.

∇S::Union{Matrix{R}, Nothing}: Gradient of the surface elevation.

∇Sy::Union{Matrix{R}, Nothing}: Gradient of the surface elevation in the y-direction.

∇Sx::Union{Matrix{R}, Nothing}: Gradient of the surface elevation in the x-direction.

Fx::Union{Matrix{R}, Nothing}: Flux in the x-direction.

Fy::Union{Matrix{R}, Nothing}: Flux in the y-direction.

Fxx::Union{Matrix{R}, Nothing}: Second derivative of flux in the x-direction.

Fyy::Union{Matrix{R}, Nothing}: Second derivative of flux in the y-direction.

V::Union{Matrix{R}, Nothing}: Velocity.

Vx::Union{Matrix{R}, Nothing}: Velocity in the x-direction.

Vy::Union{Matrix{R}, Nothing}: Velocity in the y-direction.

Γ::Union{Ref{R}, Nothing}: Basal shear stress.

MB::Union{Matrix{R}, Nothing}: Mass balance.

MB_mask::Union{AbstractArray{Bool}, Nothing}: Mask for mass balance.

MB_total::Union{Matrix{R}, Nothing}: Total mass balance.

glacier_idx::Union{Ref{I}, Nothing}: Index of the glacier.

Huginn.SIAmodel — Type

SIAmodel

An abstract type representing the Shallow Ice Approximation (SIA) models. This type is a subtype of IceflowModel and serves as a base for all SIA-specific models.

Huginn.SolverParameters — Type

SolverParameters{F <: AbstractFloat, I <: Integer}

A mutable struct that holds parameters for the solver.

Fields

solver::OrdinaryDiffEq.OrdinaryDiffEqAdaptiveAlgorithm: The algorithm used for solving differential equations.

reltol::F: The relative tolerance for the solver.

step::F: The step size for the solver.

tstops::Union{Nothing, Vector{F}}: Optional vector of time points where the solver should stop for the callbacks.

save_everystep::Bool: Flag indicating whether to save the solution at every step.

progress::Bool: Flag indicating whether to show progress during the solving process.

progress_steps::I: The number of steps between progress updates.

Huginn.SolverParameters — Method

SolverParameters(;
    solver::OrdinaryDiffEq.OrdinaryDiffEqAdaptiveAlgorithm = RDPK3Sp35(),
    reltol::Float64 = 1e-7
    )

Initialize the parameters for the numerical solver. Keyword arguments ================= - solver: solver to use from DifferentialEquations.jl - reltol: Relative tolerance for the solver

Huginn.Model — Method

function Model(; iceflow::Union{IFM, Nothing}, mass_balance::Union{MBM, Nothing} ) where {IFM <: IceflowModel, MBM <: MBmodel}

Initialize Model at Huginn level (no machine learning model).

Huginn.SIA2D! — Method

SIA2D!(dH, H, SIA2Dmodel)

Compute an in-place step of the Shallow Ice Approximation PDE in a forward model

Huginn.SIA2D — Method

SIA(H, SIA2Dmodel)

Compute a step of the Shallow Ice Approximation UDE in a forward model. Allocates memory.

Huginn.avg — Method

avg(A)

4-point average of a matrix

Huginn.avg_surface_V! — Method

avg_surface_V!(simulation::SIM) where {SIM <: Simulation}

Computes the average ice surface velocity for a given glacier evolution period based on the initial and final ice thickness states.

Huginn.avg_surface_V — Method

avg_surface_V(simulation::SIM; batch_id::Union{Nothing, I} = nothing) where {I <: Integer, SIM <: Simulation}

Computes the average ice surface velocity for a given glacier evolution period based on the initial and final ice thickness states.

Huginn.avg_x — Method

avg_x(A)

2-point average of a matrix's X axis

Huginn.avg_y — Method

avg_y(A)

2-point average of a matrix's Y axis

Huginn.batch_iceflow_PDE! — Method

batch_iceflow_PDE!(glacier_idx::I, simulation::Prediction)

Solve the Shallow Ice Approximation iceflow PDE for a given temperature series batch in-place.

Huginn.batch_iceflow_PDE — Method

batch_iceflow_PDE(glacier_idx::I, simulation::Prediction)

Solve the Shallow Ice Approximation iceflow PDE for a given temperature series batch out-of-place.

Huginn.define_callback_steps — Method

define_callback_steps(tspan::Tuple{F, F}, step::F) where {F <: AbstractFloat}

Defines the times to stop for the DiscreteCallback given a step

Huginn.diff_x — Method

diff_x(A)

2-point differential of a matrix's X axis

Huginn.diff_y — Method

diff_y(A)

2-point differential of a matrix's Y axis

Huginn.initialize_iceflow_model! — Method

function initializeiceflowmodel!(iceflowmodel::IF, glacieridx::I, glacier::AbstractGlacier, params::Sleipnir.Parameters ) where {IF <: IceflowModel, I <: Integer}

Initialize iceflow model data structures to enable in-place mutation.

Keyword arguments

- `iceflow_model`: Iceflow model used for simulation. 
- `glacier_idx`: Index of glacier.
- `glacier`: `Glacier` to provide basic initial state of the ice flow model.
- `parameters`: `Parameters` to configure some physical variables.

Huginn.initialize_iceflow_model — Method

function initializeiceflowmodel(iceflowmodel::IF, glacieridx::I, glacier::AbstractGlacier, params::Sleipnir.Parameters ) where {IF <: IceflowModel, I <: Integer}

Initialize iceflow model data structures to enable out-of-place mutation.

Keyword arguments

- `iceflow_model`: Iceflow model used for simulation. 
- `glacier_idx`: Index of glacier.
- `glacier`: `Glacier` to provide basic initial state of the ice flow model.
- `parameters`: `Parameters` to configure some physical variables.

Huginn.inn — Method

inn(A)

Access the inner part of the matrix (-2,-2)

Huginn.inn1 — Method

inn1(A)

Access the inner part of the matrix (-1,-1)

Huginn.run! — Method

run!(simulation::Prediction)

In-place run of the model.

Huginn.run₀ — Method

run(simulation::Prediction)

Out-of-place run of the model.

Huginn.simulate_iceflow_PDE! — Method

function simulate_iceflow_PDE!(
    simulation::SIM,
    cb::DiscreteCallback;
    du = SIA2D!) where {SIM <: Simulation}

Make forward simulation of the iceflow PDE determined in du.

Huginn.simulate_iceflow_PDE — Method

function simulate_iceflow_PDE(
    simulation::SIM,
    cb::DiscreteCallback;
    du = SIA2D) where {SIM <: Simulation}

Make forward simulation of the iceflow PDE determined in du.

Huginn.surface_V! — Method

surface_V!(H, B, Δx, Δy, temp, sim, A_noise, θ=[], UA_f=[])

Computes the ice surface velocity for a given glacier state

Huginn.surface_V — Method

surface_V(H, B, Δx, Δy, temp, sim, A_noise, θ=[], UA_f=[])

Computes the ice surface velocity for a given glacier state

Muninn.TImodel1 — Method

TImodel1(params::Sleipnir.Parameters;
    DDF::Float64 = 7.0/1000.0,
    acc_factor::Float64 = 1.0/1000.0
    )

Temperature-index model with a single degree-day factor.

Keyword arguments

- `DDF`: Single degree-day factor, for both snow and ice.
- `acc_factor`: Accumulation factor

Muninn.TImodel2 — Method

TImodel2(params::Parameters;
    DDF_snow::Float64 = 3.0/1000.0,
    DDF_ice::Float64 = 6.0/1000.0,
    acc_factor::Float64 = 1.0/1000.0
    )

Temperature-index model with two melt factors, for snow and ice.

Keyword arguments

- `DDF_snow`: Degree-day factor for snow.
- `DDF_ice`: Degree-day factor for ice.
- `acc_factor`: Accumulation factor

Muninn.MB_timestep! — Method

MB_timestep!(model::Model, glacier::G, step::F, t; batch_id::Union{Nothing, I} = nothing) where {I <: Integer, F <: AbstractFloat, G <: AbstractGlacier}

Retrieve the climate data, apply downscaling and compute the mass balance of a glacier for a specific time step and possibly for a batch of iceflow models.

Muninn.MB_timestep — Method

MB_timestep(model::Model, glacier::G, step::F, t::F) where {F <: AbstractFloat, G <: AbstractGlacier}

Retrieve the climate data, apply downscaling and compute the mass balance of a glacier for a specific time step.

Muninn.compute_MB — Method

compute_MB(mb_model::TImodel1, climate_2D_period::Climate2Dstep)

Compute the the mass balance given a mass balance model and a climate step.

Keyword arguments

- `mb_model`: Mass balance model.
- `climate_2D_period`: Climate step.

Sleipnir.Glacier1D — Method

function Glacier1D(; rgiid::Union{String, Nothing} = nothing, climate::Union{Climate1D, Nothing} = nothing, H₀::Union{Vector{F}, Nothing} = nothing, S::Union{Vector{F}, Nothing} = nothing, B::Union{Vector{F}, Nothing} = nothing, V::Union{Vector{F}, Nothing}= nothing, A::Union{F, Nothing} = nothing, C::Union{F, Nothing} = nothing, n::Union{F, Nothing} = nothing, w₀::Union{Vector{F}, Nothing} = nothing, λ::Union{Vector{F}, Nothing} = nothing, slope::Union{Vector{F}, Nothing} = nothing, distborder::Union{Vector{F}, Nothing} = nothing, Coords::Union{Dict{String, Vector{Float64}}, Nothing} = nothing, Δx::Union{F, Nothing} = nothing, Δy::Union{F, Nothing} = nothing, nx::Union{I, Nothing} = nothing, ny::Union{I, Nothing} = nothing ) where {F <: AbstractFloat, I <: Integer}

Constructor for empty 2D Glacier object.

Sleipnir.Glacier2D — Method

function Glacier2D(; rgiid::Union{String, Nothing} = nothing, climate::Union{Climate2D, Nothing} = nothing, H₀::Union{Matrix{F}, Nothing} = nothing, Hglathida::Union{Matrix{F}, Nothing} = nothing, S::Union{Matrix{F}, Nothing} = nothing, B::Union{Matrix{F}, Nothing} = nothing, V::Union{Matrix{F}, Nothing}= nothing, Vx::Union{Matrix{F}, Nothing}= nothing, Vy::Union{Matrix{F}, Nothing}= nothing, A::Union{F, Nothing} = nothing, C::Union{F, Nothing} = nothing, n::Union{F, Nothing} = nothing, slope::Union{Matrix{F}, Nothing} = nothing, dist_border::Union{Matrix{F}, Nothing} = nothing, Coords::Union{Dict{String, Vector{Float64}}, Nothing} = nothing, Δx::Union{F, Nothing} = nothing, Δy::Union{F, Nothing} = nothing, nx::Union{I, Nothing} = nothing, ny::Union{I, Nothing} = nothing, cenlon::Union{F, Nothing} = nothing, cenlat::Union{F, Nothing} = nothing ) where {F <: AbstractFloat, I <: Integer}

Constructor for empty 2D Glacier object.

Sleipnir.Parameters — Method

Parameters(; physical::PhysicalParameters = PhysicalParameters(), simulation::SimulationParameters = SimulationParameters() ) Initialize ODINN parameters

Keyword arguments

Sleipnir.PhysicalParameters — Method

PhysicalParameters(;
        ρ::F = 900.0,
        g::F = 9.81,
        ϵ::F = 1e-3,
        η₀::F = 1.0, 
        maxA::F = 8e-17,
        minA::F = 8.5e-20,
        maxTlaw::F = 1.0,
        minTlaw::F = -25.0,
        noise_A_magnitude::F = 5e-18
        )

Initialize the physical parameters of a model. Keyword arguments ================= - ρ: Ice density - g: Gravitational constant - ϵ: Small number - η₀: - maxA: Maximum value for A (Glen's coefficient) - minA: Minimum value for A (Glen's coefficient) - maxTlaw: Maximum value of Temperature used in simulations on fake law - minTlaw: Minimum value of Temperature used in simulations on fake law - noise_A_magnitude: Magnitude of noise added to A

Sleipnir.Results — Type

mutable struct Results{F <: AbstractFloat, I <: Int}

A mutable struct to store the results of simulations.

Fields

rgi_id::String: Identifier for the RGI (Randolph Glacier Inventory).

H::Vector{Matrix{F}}: Vector of matrices representing glacier ice thickness H over time.

H_glathida::Union{Nothing, Vector{Matrix{F}}}: Optional vector of matrices for Glathida ice thicknesses.

S::Matrix{F}: Glacier surface altimetry.

B::Matrix{F}: Glacier bedrock.

V::Matrix{F}: Glacier ice surface velocities.

Vx::Matrix{F}: x-component of the glacier ice surface velocity V.

Vy::Matrix{F}: y-component of the glacier ice surface velocity V.

V_ref::Union{Nothing, Matrix{F}}: Reference data for glacier ice surface velocities V.

Vx_ref::Union{Nothing, Matrix{F}}: Reference data for the x-component of the glacier ice surface velocity Vx.

Vy_ref::Union{Nothing, Matrix{F}}: Reference data for the y-component of the glacier ice surface velocity Vy.

Δx::F: Grid spacing in the x-direction.

Δy::F: Grid spacing in the y-direction.

lon::Union{Nothing, F}: Optional longitude value.

lat::Union{Nothing, F}: Optional latitude value.

nx::I: Number of grid points in the x-direction.

ny::I: Number of grid points in the y-direction.

tspan::Vector{F}: Time span of the simulation.

θ::Union{Nothing, ComponentArray{F}}: Machine learning model parameters.

loss::Union{Nothing, Vector{F}} Vector with evolution of loss function.

Sleipnir.Results — Method

Results(glacier::G, ifm::IF; rgi_id::String = glacier.rgi_id, H::Vector{Matrix{F}} = Vector{Matrix{F}}([]), 
        H_glathida::Union{Nothing, Vector{Matrix{F}}} = glacier.H_glathida, S::Matrix{F} = zeros(F, size(ifm.S)), 
        B::Matrix{F} = zeros(F, size(ifm.B)), V::Matrix{F} = zeros(F, size(ifm.V)), 
        Vx::Matrix{F} = zeros(F, size(ifm.Vx)), Vy::Matrix{F} = zeros(F, size(ifm.Vy)), 
        V_ref::Union{Nothing, Matrix{F}} = glacier.V, Vx_ref::Union{Nothing, Matrix{F}} = glacier.Vx, 
        Vy_ref::Union{Nothing, Matrix{F}} = glacier.Vy, Δx::F = glacier.Δx, Δy::F = glacier.Δy, 
        lon::Union{Nothing, F} = glacier.cenlon, lat::Union{Nothing, F} = glacier.cenlat, 
        nx::I = glacier.nx, ny::I = glacier.ny, θ::Union{Nothing, ComponentArray{F}} = nothing,
        loss::Union{Nothing, Vector{F}} = Nothing) where {G <: AbstractGlacier, F <: AbstractFloat, IF <: AbstractModel}

Construct a Results object for a glacier simulation.

Arguments

glacier::G: The glacier object, subtype of AbstractGlacier.

ifm::IF: The model object, subtype of AbstractModel.

rgi_id::String: The RGI identifier for the glacier. Defaults to glacier.rgi_id.

H::Vector{Matrix{F}}: Ice thickness matrices. Defaults to an empty vector.

H_glathida::Union{Nothing, Vector{Matrix{F}}}: Ice thickness from GlaThiDa. Defaults to glacier.H_glathida.

S::Matrix{F}: Surface elevation matrix. Defaults to a zero matrix of the same size as ifm.S.

B::Matrix{F}: Bed elevation matrix. Defaults to a zero matrix of the same size as ifm.B.

V::Matrix{F}: Velocity magnitude matrix. Defaults to a zero matrix of the same size as ifm.V.

Vx::Matrix{F}: Velocity in the x-direction matrix. Defaults to a zero matrix of the same size as ifm.Vx.

Vy::Matrix{F}: Velocity in the y-direction matrix. Defaults to a zero matrix of the same size as ifm.Vy.

V_ref::Union{Nothing, Matrix{F}}: Reference velocity magnitude matrix. Defaults to glacier.V.

Vx_ref::Union{Nothing, Matrix{F}}: Reference velocity in the x-direction matrix. Defaults to glacier.Vx.

Vy_ref::Union{Nothing, Matrix{F}}: Reference velocity in the y-direction matrix. Defaults to glacier.Vy.

Δx::F: Grid spacing in the x-direction. Defaults to glacier.Δx.

Δy::F: Grid spacing in the y-direction. Defaults to glacier.Δy.

lon::Union{Nothing, F}: Longitude of the glacier center. Defaults to glacier.cenlon.

lat::Union{Nothing, F}: Latitude of the glacier center. Defaults to glacier.cenlat.

nx::I: Number of grid points in the x-direction. Defaults to glacier.nx.

ny::I: Number of grid points in the y-direction. Defaults to glacier.ny.

tspan::Tuple(F, F): Timespan of the simulation.

θ::Union{Nothing, ComponentArray{F}}: Model parameters. Defaults to nothing.

loss::Union{Nothing, Vector{F}}: Loss values. Defaults to nothing.

Returns

results::Results: A Results object containing the simulation results.

Sleipnir.SimulationParameters — Method

SimulationParameters(;
                    use_MB::Bool = true,
                    use_iceflow::Bool = true,
                    plots::Bool = true,
                    velocities::Bool = true,
                    overwrite_climate::Bool = false,
                    use_glathida_data::Bool = false,
                    tspan::Tuple{F, F} = (2010.0,2015.0),
                    step::F = 1/12,
                    multiprocessing::Bool = true,
                    workers::I = 4,
                    working_dir::String = "",
                    test_mode::Bool = false,
                    rgi_paths::Dict{String, String} = Dict{String, String}(),
                    ice_thickness_source::String = "Farinotti19",
    )

Initialize the parameters for a simulation. Keyword arguments ================= - use_MB: Determines if surface mass balance should be used. - plots: Determines if plots should be made. - overwrite_climate: Determines if climate data should be overwritten - 'useglathidadata': Determines if data from the Glathida data set should be used

Sleipnir.SurfaceVelocityData — Method

function SurfaceVelocityData(; x::Union{Vector{F}, Nothing} = nothing, y::Union{Vector{F}, Nothing} = nothing, lat::Union{Vector{F}, Nothing} = nothing, lon::Union{Vector{F}, Nothing} = nothing, vx::Union{Array{F, 3}, Nothing} = nothing, vy::Union{Array{F, 3}, Nothing} = nothing, vabs::Union{Array{F, 3}, Nothing} = nothing, vxerror::Union{Array{F, 1}, Nothing} = nothing, vyerror::Union{Array{F, 1}, Nothing} = nothing, vabserror::Union{Array{F, 1}, Nothing} = nothing, date::Union{Vector{DateTime}, Nothing} = nothing, date1::Union{Vector{DateTime}, Nothing} = nothing, date2::Union{Vector{DateTime}, Nothing} = nothing, dateerror::Union{Vector{Day}, Vector{Millisecond}, Nothing} = nothing, ) where {F <: AbstractFloat}

Constructor for ice surface velocity data based on Rabatel et. al (2023).

Important remarks:

Projections in longitude and latitude assume we are working in the north hemisphere. If working with south hemisphere glaciers, this needs to be changed.

The error in velocity is unique per timestamp, rather than being pixel distributed.

The error in the absolute velocities vabs_error is overestimated.

References: - Rabatel, A., Ducasse, E., Millan, R. & Mouginot, J. Satellite-Derived Annual Glacier Surface Flow Velocity Products for the European Alps, 2015–2021. Data 8, 66 (2023).

Sleipnir.UTMercator — Method

UTMercator(x::F, y::F; k=0.9996, cenlon=0.0, cenlat=0.0, x0=0.0, y0=0.0)

Transverse Mercator Projection. This function reprojects northing/easting coordinates into latitude/longitude.

Keyword arguments

- `k`: scale factor of the projection
- `cenlon`: Central longitude used in the projection
- `cenlat`: Central latitude used in the projection
- `x0`: Shift in easting
- `y0`: Shift in northing
- `zone` : Zone of the projection
- `hemisphere`: Either :north or :south

Sleipnir.apply_t_grad! — Method

apply_t_grad!(climate::RasterStack, dem::Raster)

Applies temperature gradients to the glacier 2D climate data based on a DEM.

Sleipnir.create_results — Method

store_results!(simulation::SIM, glacier_idx::I, solution) where {SIM <: Simulation, I <: Int}

Store the results of a simulation of a single glacier into a Results.

Sleipnir.downscale_2D_climate! — Method

downscale_2D_climate(glacier::Glacier2D)

Projects climate data to the glacier matrix by simply copying the closest gridpoint to all matrix gridpoints. Generates a new RasterStack which is returned.

Sleipnir.fillNaN! — Function

fillNaN!(x, fill)

Convert empty matrix grid cells into fill value

Sleipnir.get_cumulative_climate! — Function

get_cumulative_climate(climate, gradient_bounds=[-0.009, -0.003], default_grad=-0.0065)

Computes Positive Degree Days (PDDs) and cumulative rainfall and snowfall from climate data.

Sleipnir.get_raw_climate_data — Method

get_raw_climate_data(rgi_path::String)

Load the netCDF file containing the climate data for that glacier.

Sleipnir.initialize_glacier — Method

initialize_glacier(rgi_id::String, parameters::Parameters; smoothing=false, velocities=true)

Initialize a single Glaciers, including its Climate, based on a rgi_id and timestepping arguments.

Keyword arguments

- `rgi_id`: Glacier RGI ID
- `parameters`: Parameters including the physical and simulation ones
- `smoothing` Flag determining if smoothing needs to be applied to the surface elevation and ice thickness.
- `velocities` Flag determining if the ice surface velocities need to be retrieved.

Sleipnir.initialize_glacier_climate! — Method

function initialize_glacier_climate!(glacier::AbstractGlacier, params::Parameters)

Initializes the Climate data structure for a given Glacier`

Sleipnir.initialize_glacier_data — Method

initialize_glacier(rgi_id::String, params::Parameters; smoothing=false, velocities=true)

Retrieves the initial glacier geometry (bedrock + ice thickness) for a glacier with other necessary data (e.g. grid size and ice surface velocities).

Sleipnir.initialize_glaciers — Method

initialize_glaciers(rgi_ids::Vector{String}, params::Parameters; test=false)

Initialize multiple Glaciers based on a list of RGI IDs and on parameters.

Keyword arguments

- `rgi_ids`: List of RGI IDs of glaciers
- `params`: `Parameters` object to be passed

Sleipnir.initialize_surfacevelocitydata — Method

initialize_surfacevelocitydata(file::String; interp=false)

Initialize SurfaceVelocityData from Rabatel et. al (2023).

Arguments

- `file`: name of netcdf file with data
- `interp`: boolean variable indicating if we use the inporpolated data or not

Sleipnir.max_or_empty — Method

max_or_empyt(A::Array)

Return maximum value for non-empty arrays. This is just required to compute the error in the absolute velocity.

Sleipnir.parse_proj — Method

parse_proj(proj::String)

Parses the string containing the information of the projection to filter for important information "+proj=tmerc +lat0=0 +lon0=6.985 +k=0.9996 +x0=0 +y0=0 +datum=WGS84 +units=m +no_defs"

Sleipnir.plot_bias — Method

plot_bias(data::Results, variables::Vector{Symbol}; treshold::Vector{Float64}=[0, 0])

Plot the bias of the glacier integrated volume over the specified time span.

Arguments

results::Results: The results object containing the data to be plotted.

variables::Vector{Symbol}: The variables to be plotted.

title_mapping::Dict{Symbol, String}: A dictionary mapping variable names to their titles.

tspan::Tuple{Float64, Float64}: A tuple representing the start and end time for the simulation.

Returns

A plot of the glacier integrated volume bias.

Sleipnir.plot_glacier — Method

plot_glacier(results::Results, plot_type::String, variables::Vector{Symbol}; kwargs...) -> Figure

Generate various types of plots for glacier data.

Arguments

results::Results: The results object containing the data to be plotted.

plot_type::String: Type of plot to generate. Options are:

"heatmaps": Heatmaps for glacier variables like :H, :S, :B, :V, :Vx, and :Vy.

"evolution difference": Temporal difference metrics (between start and end) for a variable, with optional metrics like "hist" (histogram) and "difference".

"evolution statistics": Temporal statistical metrics for a variable, with optional metrics like "average", "median", "min", "max", and "std".

"integrated volume": Temporal evolution of the integrated ice volume for a variable.

"bias": Scatter plot to visualize the bias between two variables.

variables::Vector{Symbol}: Variables to be plotted, e.g., :H.

Optional Keyword Arguments

tspan: A tuple representing the start and end time for the simulation.

metrics: Metrics to visualize, e.g., ["average"] for statistics, ["difference"] for difference.

scale_text_size::Union{Nothing,Float64}: Optional argument to scale the text size for heatmaps.

threshold::Vector{F}: Threshold values for filtering data in bias plots.

Returns

A Figure object containing the desired visualization.

Notes

Ensure the variables and kwargs match the requirements of the specified plot_type.

The function routes requests to specific plotting functions based on plot_type.

Sleipnir.plot_glacier_difference_evolution — Method

plot_glacier_difference_evolution(results::Results, variables::Vector{Symbol}, title_mapping; tspan::Tuple{F,F}=results.tspan, metrics::Vector{String}="difference") where {F<:AbstractFloat}

Plot the evolution of the difference in a glacier variable over time.

Arguments

results::Results: The simulation results object containing the data to be plotted.

variables::Vector{Symbol}: The variable to be plotted.

title_mapping: A dictionary mapping variable names to their titles.

tspan::Tuple{F,F}: A tuple representing the start and end time for the simulation.

metrics::Vector{String}: Metrics to visualize, e.g., ["difference"].

Returns

A plot of the glacier difference evolution.

Sleipnir.plot_glacier_heatmaps — Method

plot_glacier_heatmaps(results::Results, variables::Vector{Symbol}, title_mapping::Dict; scale_text_size::Union{Nothing,Float64}=nothing) -> Figure

Plot heatmaps for glacier variables.

Arguments

results::Results: The results object containing the data to be plotted.

variables::Vector{Symbol}: A list of variables to be plotted.

title_mapping::Dict: A dictionary mapping variable names to their titles and colormaps.

scale_text_size::Union{Nothing,Float64}: Optional argument to scale the text size.

Returns

A plot of the glacier heatmaps.

Sleipnir.plot_glacier_integrated_volume — Method

plot_glacier_integrated_volume(results::Results, variables::Vector{Symbol}, title_mapping; tspan)

Plot the integrated volume of a glacier variable over time.

Arguments

results::Results: The results object containing the data to be plotted.

variables::Vector{Symbol}: The variable to be plotted.

title_mapping: A dictionary mapping variable names to their titles.

tspan: A tuple representing the start and end time for the simulation.

Returns

A plot of the glacier integrated volume.

Sleipnir.plot_glacier_statistics_evolution — Method

plot_glacier_statistics_evolution(results::Results, variables::Vector{Symbol}, title_mapping; metrics="median", tspan, threshold=0.5)

Plot the evolution of statistics for multiple glacier variables over time.

Arguments

results::Results: The simulation results object containing the data to be plotted.

variables::Vector{Symbol}: A list of variables to be plotted.

title_mapping: A dictionary mapping variable names to their titles.

metrics: Metrics to visualize, e.g., "average", "median", "min", "max", and "std". Default is "median".

tspan: A tuple representing the start and end time for the simulation.

threshold: A threshold value to filter the data. Default is 0.5.

Returns

A plot of the glacier statistics evolution.

Sleipnir.plot_glacier_vid — Method

plot_glacier_vid(
    plot_type::String,
    H::Vector{Matrix{Float64}},
    glacier::Glacier2D,
    simuparams::SimulationParameters,
    pathVideo::String;
    framerate::Int=24,
    baseTitle::String=""
)

Generate various types of videos for glacier data.

Arguments

plot_type: Type of plot to generate. Options are:

"thickness": Heatmap of the glacier thickness.

H: A vector of matrices containing the ice thickness over time. This should be replaced by a Results instance in the future once Results no longer depends on an iceflow model.

glacier: A glacier instance.

simuparams: The simulation parameters.

pathVideo: Path of the mp4 file to generate.

Optional Keyword Arguments

framerate: The framerate to use for the video generation.

baseTitle: The prefix to use in the title of the frames. In each frame it is concatenated with the value of the year in the form " (t=XXXX)".

Sleipnir.save_results_file! — Method

save_results_file!(results_list::Vector{Results{F}}, simulation::SIM; path::Union{String,Nothing}=nothing) where {F <: AbstractFloat, SIM <: Simulation}

Save simulation results which are provided as a list of Results into a .jld2 file. This function also overrides the resultsattribute ofsimulation`.

Sleipnir.smooth! — Method

smooth!(A)

Smooth data contained in a matrix with one time step (CFL) of diffusion.

Sleipnir.stop_condition_tstops — Method

stop_condition_tstops(u,t,integrator, tstops)

Function that iterates through the tstops, with a closure including tstops

Sleipnir.trim_period — Method

trim_period(period, climate)

Trims a time period based on the time range of a climate series.