Index of functions and types

Parameters

There are different types of parameters, holding specific information for different modelling aspects. All the types of parameters are wrapped into a Parameter type, which is threaded throughout ODINN.jl.

Missing docstring.

Missing docstring for Sleipnir.SimulationParameters. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Sleipnir.PhysicalParameters. Check Documenter's build log for details.

ODINN.Hyperparameters — Type

Hyperparameters(;
    current_epoch::Int64 = nothing,
    current_minibatch::Int64 = nothing,
    loss_history::Vector{Float64} = Vector{Float64}(),
    optimizer::Optim.FirstOrderOptimizer = BFGS(initial_stepnorm=0.001),
    epochs::Int64 = 50,
    batch_size::Int64 = 15
    )

Initialize the hyperparameters of a machine learning model (Machine). Keyword arguments ================= - current_epoch: Current epoch in training - current_minibatch: Current minibatch in training - loss_history: Vector storing the loss for each epoch during training - optimizer: Optimizer to be used for training. Currently supports both Optim.jl and Flux.jl optimisers. - epochs: Number of epochs for the training - batch_size: Batch size for the training

ODINN.UDEparameters — Type

UDEparameters(;
    sensealg::SciMLBase.AbstractAdjointSensitivityAlgorithm = GaussAdjoint(autojacvec=SciMLSensitivity.EnzymeVJP()),
    optimization_method::String = "AD+AD",
    loss_type::String = "V",
    empirical_loss_function::AbstractLoss = L2Sum(),
    scale_loss::Bool = true
    target::Union{String, Nothing} = "D"
    )

Initialize the parameters for the training of the UDE. Keyword arguments ================= - sensealg: Sensitivity algorithm from SciMLSensitivity.jl to be used. - optimization_method: Optimization method for the UDE. - loss_type: Type of loss function to be used. Can be either V (ice velocities), or H (ice thickness). - empirical_loss_function: Empirical loss function to use. - scale_loss: Determines if the loss function should be scaled or not.

All these subtypes of parameters are gathered in a Parameters struct:

ODINN.Parameters — Function

Parameters(; physical::PhysicalParameters = PhysicalParameters(), simulation::SimulationParameters = SimulationParameters(), solver::SolverParameters = SolverParameters(), hyper::Hyperparameters = Hyperparameters(), UDE::UDEparameters = UDEparameters() inversion::InversionParameters = InversionParameters() ) Initialize ODINN parameters

Keyword arguments

physical::PhysicalParameters: Physical parameters for the simulation.

simulation::SimulationParameters: Parameters related to the simulation setup.

solver::SolverParameters: Parameters for the solver configuration.

hyper::Hyperparameters: Hyperparameters for the model.

UDE::UDEparameters: Parameters specific to the UDE (Universal Differential Equation).

inversion::InversionParameters: Parameters for inversion processes.

Glaciers

Glaciers in ODINN.jl are represented by a Glacier type. Each glacier has its related Climate. Since ODINN.jl supports different types of simulations, we offer the possibility to work on 1D (i.e. flowline), 2D (e.g. SIA) or even 3D (not yet implemented, e.g. Full Stokes).

Missing docstring.

Missing docstring for Sleipnir.Glacier2D. Check Documenter's build log for details.

Every glacier has its associated climate, following the same spatial representation (e.g. 2D):

<!– @docs Sleipnir.Climate2D –>

Models

There are 3 main types of models in ODINN.jl, iceflow models, mass balance models and machine learning models.

Work in progress...