Types and functions

In this page, we will go through the main types (i.e. structs) used in ODINN.jl's architecture, and the main functions linked to those types.

Parameters

There are different types of parameters, holding specific information for different modelling aspects. All the types of parameters are wrapped into a Parameter type, which is threaded throughout ODINN.jl.

Sleipnir.Parameters — Type

    mutable struct Parameters{PPHY <: AbstractEmptyParams, PSIM <: AbstractEmptyParams, PHY <: AbstractEmptyParams,
                    PSOL <: AbstractEmptyParams, PUDE <: AbstractEmptyParams, PINV <: AbstractEmptyParams}

A mutable struct that holds various parameter sets for different aspects of a simulation or model.

Fields

physical::PPHY: Physical parameters.

simulation::PSIM: Simulation parameters.

hyper::PHY: Hyperparameters.

solver::PSOL: Solver parameters.

UDE::PUDE: Universal Differential Equation (UDE) parameters.

inversion::PINV: Inversion parameters.

Type Parameters

PPHY: Type of the physical parameters, must be a subtype of AbstractEmptyParams.

PSIM: Type of the simulation parameters, must be a subtype of AbstractEmptyParams.

PHY: Type of the hyperparameters, must be a subtype of AbstractEmptyParams.

PSOL: Type of the solver parameters, must be a subtype of AbstractEmptyParams.

PUDE: Type of the UDE parameters, must be a subtype of AbstractEmptyParams.

PINV: Type of the inversion parameters, must be a subtype of AbstractEmptyParams.

ODINN.Parameters — Function

Constructor for the Parameters type. Since some of the subtypes of parameters are defined in different packages of the ODINN ecosystem, this constructor will call the constructors of the different subtypes and return a Parameters object with the corresponding subtypes. The Parameters mutable struct is defined in Sleipnir.jl using abstract types, which are later on defined in the different packages of the ODINN ecosystem.

Parameters(;
        physical::PhysicalParameters = PhysicalParameters(),
        simulation::SimulationParameters = SimulationParameters(),
        solver::SolverParameters = SolverParameters(),
        hyper::Hyperparameters = Hyperparameters(),
        UDE::UDEparameters = UDEparameters()
        inversion::InversionParameters = InversionParameters()
        )

Keyword arguments

physical::PhysicalParameters: Physical parameters for the simulation.

simulation::SimulationParameters: Parameters related to the simulation setup.

solver::SolverParameters: Parameters for the solver configuration.

hyper::Hyperparameters: Hyperparameters for the model.

UDE::UDEparameters: Parameters specific to the UDE (Universal Differential Equation).

inversion::InversionParameters: Parameters for inversion processes.

The main types of parameters are the following ones:

Simulation parameters

Simulation parameters are used to specify anything related to ODINN simulations, ranging from types, working directories to multiprocessing.

Sleipnir.SimulationParameters — Type

A structure to hold simulation parameters for a simulation in ODINN.

struct SimulationParameters{I <: Integer, F <: AbstractFloat, VM <: VelocityMapping} <: AbstractParameters

Fields

use_MB::Bool: Flag to indicate whether mass balance should be used.

use_iceflow::Bool: Flag to indicate whether ice flow should be used.

plots::Bool: Flag to indicate whether plots should be generated.

velocities::Bool: Flag to indicate whether velocities should be calculated.

overwrite_climate::Bool: Flag to indicate whether to overwrite climate data.

use_glathida_data::Bool: Flag to indicate whether to use GLATHIDA data.

tspan::Tuple{F, F}: Time span for the simulation.

step::F: Time step for the simulation.

multiprocessing::Bool: Flag to indicate whether multiprocessing should be used.

workers::I: Number of workers for multiprocessing.

working_dir::String: Directory for working files.

test_mode::Bool: Flag to indicate whether to run in test mode.

rgi_paths::Dict{String, String}: Dictionary of RGI paths.

ice_thickness_source::String: Source of ice thickness data.

mapping::VM: Mapping to use in order to grid the data from the coordinates of the velocity product datacube to the glacier grid.

gridScalingFactor::I: Grid downscaling factor, used to speed-up the tests. Default value is 1 which means no downscaling is applied.

Sleipnir.SimulationParameters — Method

Constructor for SimulationParameters type, including default values.

SimulationParameters(;
    use_MB::Bool = true,
    use_iceflow::Bool = true,
    plots::Bool = true,
    velocities::Bool = true,
    overwrite_climate::Bool = false,
    use_glathida_data::Bool = false,
    tspan::Tuple{F, F} = (2010.0,2015.0),
    step::F = 1/12,
    multiprocessing::Bool = true,
    workers::I = 4,
    working_dir::String = "",
    test_mode::Bool = false,
    rgi_paths::Dict{String, String} = Dict{String, String}(),
    ice_thickness_source::String = "Farinotti19",
    mapping::VM = MeanDateVelocityMapping(),
    gridScalingFactor::I = 1,
) where {I <: Integer, F <: AbstractFloat, VM <: VelocityMapping}

Keyword arguments

use_MB::Bool: Whether to use mass balance (default: true).

use_iceflow::Bool: Whether to use ice flow (default: true).

plots::Bool: Whether to generate plots (default: true).

velocities::Bool: Whether to calculate velocities (default: true).

overwrite_climate::Bool: Whether to overwrite climate data (default: false).

use_glathida_data::Bool: Whether to use GLATHIDA data (default: false).

float_type::DataType: Data type for floating point numbers (default: Float64).

int_type::DataType: Data type for integers (default: Int64).

tspan::Tuple{F, F}: Time span for the simulation (default: (2010.0, 2015.0)).

step::F: Time step for the simulation (default: 1/12).

multiprocessing::Bool: Whether to use multiprocessing (default: true).

workers::I: Number of workers for multiprocessing (default: 4).

working_dir::String: Working directory for the simulation (default: "").

test_mode::Bool: Whether to run in test mode (default: false).

rgi_paths::Dict{String, String}: Dictionary of RGI paths (default: Dict{String, String}()).

ice_thickness_source::String: Source of ice thickness data, either "Millan22" or "Farinotti19" (default: "Farinotti19").

mapping::VM: Mapping to use in order to grid the data from the coordinates of the velocity product datacube to the glacier grid.

gridScalingFactor::I: Grid downscaling factor, used to speed-up the tests. Default value is 1 which means no downscaling is applied.

Returns

simulation_parameters: A new SimulationParameters object.

Throws

AssertionError: If ice_thickness_source is not "Millan22" or "Farinotti19".

Notes

If the global variable ODINNOVERWRITEMULTI is set to true, multiprocessing is disabled in any case. This is to fix the documentation generation as for the moment Literate.jl freezes when multiprocessing is enabled.

Physical parameters

Physical parameters are used to store physical constants used in the physical and machine learning models.

Sleipnir.PhysicalParameters — Type

A structure representing physical parameters used in simulations.

PhysicalParameters{F <: AbstractFloat}

Fields

ρ::F: Density of ice.

g::F: Gravitational acceleration.

ϵ::F: A small parameter, often used for perturbations.

η₀::F: Initial viscosity.

maxA::F: Maximum A.

minA::F: Minimum A.

maxTlaw::F: Maximum temperature according to some law.

minTlaw::F: Minimum temperature according to some law.

noise_A_magnitude::F: Magnitude of noise in A.

Sleipnir.PhysicalParameters — Method

Initialize the physical parameters of a model.

PhysicalParameters(;
    ρ::Float64 = 900.0,
    g::Float64 = 9.81,
    ϵ::Float64 = 1e-3,
    η₀::F = 1.0, 
    maxA::Float64 = 8e-17,
    minA::Float64 = 8.5e-20,
    maxTlaw::Float64 = 1.0,
    minTlaw::Float64 = -25.0,
    noise_A_magnitude::Float64 = 5e-18
    )

Keyword arguments

- `ρ`: Ice density
- `g`: Gravitational constant
- `ϵ`: Small number
- `η₀`:  
- `maxA`: Maximum value for `A` (Glen's coefficient)
- `minA`: Minimum value for `A` (Glen's coefficient)
- `maxTlaw`: Maximum value of Temperature used in simulations on fake law
- `minTlaw`: Minimum value of Temperature used in simulations on fake law
- `noise_A_magnitude`: Magnitude of noise added to A

Solver parameters

Solver parameters determine all aspects related to the numerical scheme used to solve the differential equations of glacier ice flow.

Huginn.SolverParameters — Type

A mutable struct that holds parameters for the solver.

SolverParameters{F <: AbstractFloat, I <: Integer}

Fields

solver::OrdinaryDiffEq.OrdinaryDiffEqAdaptiveAlgorithm: The algorithm used for solving differential equations.

reltol::F: The relative tolerance for the solver.

step::F: The step size for the solver.

tstops::Union{Nothing, Vector{F}}: Optional vector of time points where the solver should stop for the callbacks.

save_everystep::Bool: Flag indicating whether to save the solution at every step.

progress::Bool: Flag indicating whether to show progress during the solving process.

progress_steps::I: The number of steps between progress updates.

Huginn.SolverParameters — Method

Constructs a SolverParameters object with the specified parameters or using default values.

SolverParameters(; solver::OrdinaryDiffEq.OrdinaryDiffEqAdaptiveAlgorithm = RDPK3Sp35(),
                  reltol::F = 1e-12,
                  step::F = 1.0/12.0,
                  tstops::Union{Nothing,Vector{F}} = nothing,
                  save_everystep = false,
                  progress::Bool = true,
                  progress_steps::I = 10) where {F <: AbstractFloat, I <: Integer}

Arguments

solver::OrdinaryDiffEq.OrdinaryDiffEqAdaptiveAlgorithm: The ODE solver algorithm to use. Defaults to RDPK3Sp35().

reltol::F: The relative tolerance for the solver. Defaults to 1e-12.

step::F: The step size for the callbacks. These are mainly used to run the surface mass balance model. Defaults to 1.0/12.0 (i.e. a month).

tstops::Union{Nothing, Vector{F}}: Optional vector of time points where the solver should stop. Defaults to nothing.

save_everystep::Bool: Whether to save the solution at every step. Defaults to false.

progress::Bool: Whether to show progress during the solving process. Defaults to true.

progress_steps::I: The number of steps between progress updates. Defaults to 10.

Returns

solver_parameters: A SolverParameters object constructed with the specified parameters.

Hyperparameters

Hyperparameters determine different aspects of a given machine learning model. For now, these are focused on neural networks, but we plan to extend them in the future for other types of regressors.

ODINN.Hyperparameters — Type

mutable struct Hyperparameters{F <: AbstractFloat, I <: Int} <: AbstractParameters

A mutable struct that holds hyperparameters for training a machine learning model.

Keyword arguments

current_epoch::I: The current epoch number.

current_minibatch::I: The current minibatch number.

loss_history::Vector{F}: A vector storing the history of loss values.

optimizer::Union{Optim.FirstOrderOptimizer, Flux.Optimise.AbstractOptimiser, Optimisers.AbstractRule}: The optimizer used for training.

loss_epoch::F: The loss value for the current epoch.

epochs::I: The total number of epochs for training.

batch_size::I: The size of each minibatch.

ODINN.Hyperparameters — Method

Hyperparameters(; current_epoch::Int64 = 1, current_minibatch::Int64 = 1, loss_history::Vector{Float64} = Vector{Float64}(), optimizer::Union{Optim.FirstOrderOptimizer, Flux.Optimise.AbstractOptimiser, Optimisers.AbstractRule} = BFGS(initial_stepnorm=0.001), loss_epoch::Float64 = 0.0, epochs::Int64 = 50, batch_size::Int64 = 15)

Constructs a Hyperparameters object with the specified parameters.

Arguments

current_epoch::Int64: The current epoch number. Defaults to 1.

current_minibatch::Int64: The current minibatch number. Defaults to 1.

loss_history::Vector{Float64}: A vector to store the history of loss values. Defaults to an empty vector.

optimizer::Union{Optim.FirstOrderOptimizer, Flux.Optimise.AbstractOptimiser, Optimisers.AbstractRule}: The optimizer to be used. Defaults to BFGS(initial_stepnorm=0.001).

loss_epoch::Float64: The loss value for the current epoch. Defaults to 0.0.

epochs::Int64: The total number of epochs. Defaults to 50.

batch_size::Int64: The size of each minibatch. Defaults to 15.

Returns

A Hyperparameters object initialized with the provided values.

UDE parameters

Universal Differential Equation (UDE) parameters are used to determine different modelling choices regarding the use of UDEs, such as wich sensitivity algorithm to use, which target (e.g. SIA parameter to target), or which optimization method to use.

ODINN.UDEparameters — Type

A mutable struct that holds parameters for a UDE (Universal Differential Equation).

UDEparameters{ADJ <: AbstractAdjointMethod} <: AbstractParameters

Fields

sensealg::SciMLBase.AbstractAdjointSensitivityAlgorithm: The sensitivity algorithm used for adjoint sensitivity analysis.

optimization_method::String: The optimization method to be used.

loss_type::String: The type of loss function to be used.

scale_loss::Bool: A boolean indicating whether to scale the loss.

target::Symbol: The target variable for the optimization.

ODINN.UDEparameters — Method

UDEparameters(; sensealg, optim_autoAD, grad, optimization_method, loss_type, empirical_loss_function, scale_loss, target) where {ADJ <: AbstractAdjointMethod}

Create a UDEparameters object for configuring the sensitivity analysis and optimization of a Universal Differential Equation (UDE).

Keyword Arguments

sensealg::SciMLBase.AbstractAdjointSensitivityAlgorithm: The sensitivity algorithm to use for adjoint calculations. Defaults to GaussAdjoint(autojacvec=SciMLSensitivity.EnzymeVJP()).

optim_autoAD::AbstractADType: The automatic differentiation type for optimization. Defaults to Optimization.AutoEnzyme().

grad::ADJ: The adjoint gradient computation method. Defaults to SciMLSensitivityAdjoint().

optimization_method::String: The optimization method to use. Must be either "AD+AD" or "AD+Diff". Defaults to "AD+AD".

loss_type::String: The type of loss function to use. Must be either "V" (velocity) or "H" (thickness). Defaults to "V".

empirical_loss_function::AbstractLoss: The loss function to use for optimization. Defaults to L2Sum().

scale_loss::Bool: Whether to scale the loss function. Defaults to true.

target::Union{Symbol, Nothing}: The target variable for optimization. Defaults to :A.

Returns

A UDEparameters object configured with the specified sensitivity, optimization, and loss settings.

Description

This function creates a UDEparameters object that encapsulates the configuration for sensitivity analysis, optimization, and loss computation in a Universal Differential Equation (UDE) framework. It verifies that the provided optimization_method and loss_type are valid and constructs the solver parameters accordingly.

Notes

The optimization_method must be either "AD+AD" (automatic differentiation for both forward and backward passes) or "AD+Diff" (automatic differentiation combined with finite differences).

The loss_type must be either "V" (velocity-based loss) or "H" (thickness-based loss).

The empirical_loss_function determines how the loss is computed during optimization.

Glaciers

Glaciers in ODINN.jl are represented by a Glacier type. Each glacier has its related Climate type. Since ODINN.jl supports different types of simulations, we offer the possibility to work on 1D (i.e. flowline), 2D (e.g. SIA) or even 3D (not yet implemented, e.g. Full Stokes).

Sleipnir.Glacier2D — Type

A mutable struct representing a 2D glacier. Notice that all fields can be empty by providing nothing as the default value.

/!\ WARNING /!\ Glacier objects should not be constructed manually, but rather through the initialize_glaciers function.

Glacier2D{F <: AbstractFloat, I <: Integer}

Fields

rgi_id::String: The RGI (Randolph Glacier Inventory) identifier for the glacier.

name::String: The name of the glacier if available.

climate::Union{Climate2D, Nothing}: The climate data associated with the glacier.

H₀::Union{Matrix{F}, Nothing}: Initial ice thickness matrix.

H_glathida::Matrix{F}: Ice thickness matrix from the GLATHIDA dataset.

S::Matrix{F}: Surface elevation matrix.

B::Matrix{F}: Bedrock elevation matrix.

V::Union{Matrix{F}, Nothing}: Ice velocity magnitude matrix.

Vx::Union{Matrix{F}, Nothing}: Ice velocity in the x-direction matrix.

Vy::Union{Matrix{F}, Nothing}: Ice velocity in the y-direction matrix.

A::Union{F, Nothing}: Flow law parameter.

C::Union{F, Nothing}: Sliding law parameter.

n::Union{F, Nothing}: Flow law exponent.

slope::Union{Matrix{F}, Nothing}: Surface slope matrix.

dist_border::Union{Matrix{F}, Nothing}: Distance to the glacier border matrix.

Coords::Union{Dict{String, Vector{Float64}}, Nothing}: Coordinates dictionary with keys as coordinate names and values as vectors of coordinates.

Δx::F: Grid spacing in the x-direction.

Δy::F: Grid spacing in the y-direction.

nx::I: Number of grid points in the x-direction.

ny::I: Number of grid points in the y-direction.

cenlon::Union{F, Nothing}: Longitude of the glacier center.

cenlat::Union{F, Nothing}: Latitude of the glacier center.

params_projection::Dict{String, Float64}: Projection parameters that allows mapping the regional grid to global WGS84 coordinates.

thicknessData::Union{ThicknessData, Nothing}: Thickness data structure that is used to store the reference values.

velocityData::Union{SurfaceVelocityData, Nothing}: Surface velocity data structure that is used to store the reference values.

Sleipnir.Glacier2D — Method

Constructs a Glacier2D object with the given parameters, including default ones.

Glacier2D(;
    rgi_id::Union{String, Nothing} = nothing,
    name::String = "",
    climate::Union{Climate2D, Nothing} = nothing,
    H₀::Union{Matrix{F}, Nothing} = nothing,
    H_glathida::Union{Matrix{F}, Nothing} = nothing,
    S::Union{Matrix{F}, Nothing} = nothing,
    B::Union{Matrix{F}, Nothing} = nothing,
    V::Union{Matrix{F}, Nothing} = nothing,
    Vx::Union{Matrix{F}, Nothing} = nothing,
    Vy::Union{Matrix{F}, Nothing} = nothing,
    A::Union{F, Nothing} = nothing,
    C::Union{F, Nothing} = nothing,
    n::Union{F, Nothing} = nothing,
    slope::Union{Matrix{F}, Nothing} = nothing,
    dist_border::Union{Matrix{F}, Nothing} = nothing,
    Coords::Union{Dict{String, Vector{Float64}}, Nothing} = nothing,
    Δx::Union{F, Nothing} = nothing,
    Δy::Union{F, Nothing} = nothing,
    nx::Union{I, Nothing} = nothing,
    ny::Union{I, Nothing} = nothing,
    cenlon::Union{F, Nothing} = nothing,
    cenlat::Union{F, Nothing} = nothing,
    params_projection::Dict{String, Float64} = Dict{String, Float64}(),
    thicknessData::Union{ThicknessData, Nothing} = nothing,
    velocityData::Union{SurfaceVelocityData, Nothing} = nothing,
) where {F <: AbstractFloat, I <: Integer}

Arguments

rgi_id::String: The RGI identifier for the glacier.

name::String: The name of the glacier if available.

climate::Union{Climate2D, Nothing}: The climate data associated with the glacier.

H₀::Union{Matrix{F}, Nothing}: Initial ice thickness matrix.

H_glathida::Matrix{F}: Ice thickness matrix from GLATHIDA.

S::Matrix{F}: Surface elevation matrix.

B::Matrix{F}: Bed elevation matrix.

V::Union{Matrix{F}, Nothing}: Ice velocity magnitude matrix.

Vx::Union{Matrix{F}, Nothing}: Ice velocity in the x-direction matrix.

Vy::Union{Matrix{F}, Nothing}: Ice velocity in the y-direction matrix.

A::Union{F, Nothing}: Flow law parameter.

C::Union{F, Nothing}: Sliding law parameter.

n::Union{F, Nothing}: Flow law exponent.

slope::Union{Matrix{F}, Nothing}: Slope matrix.

dist_border::Union{Matrix{F}, Nothing}: Distance to border matrix.

Coords::Union{Dict{String, Vector{Float64}}, Nothing}: Coordinates dictionary with keys "lon" and "lat".

Δx::F: Grid spacing in the x-direction.

Δy::F: Grid spacing in the y-direction.

nx::I: Number of grid points in the x-direction.

ny::I: Number of grid points in the y-direction.

cenlon::Union{F, Nothing}: Central longitude of the glacier.

cenlat::Union{F, Nothing}: Central latitude of the glacier.

params_projection::Dict{String, Float64}: Projection parameters that allows mapping the regional grid to global WGS84 coordinates.

thicknessData::Union{ThicknessData, Nothing}: Thickness data structure that is used to store the reference values.

velocityData::Union{SurfaceVelocityData, Nothing}: Surface velocity data structure that is used to store the reference values.

Returns

A Glacier2D object with the specified parameters.

Every glacier has its associated climate, following the same spatial representation (e.g. 2D):

Sleipnir.Climate2D — Type

A mutable struct representing a 2D climate for a glacier with various buffers and datasets.

Climate2D{F <: AbstractFloat}

Keyword arguments

raw_climate::RasterStack: Raw climate dataset for the whole simulation.

climate_raw_step::RasterStack: Raw climate trimmed for the current step to avoid memory allocations.

climate_step::Dict: Climate data for the current step.

climate_2D_step::Climate2Dstep: 2D climate data for the current step to feed to the mass balance (MB) model.

longterm_temps::Vector{F}: Long-term temperatures for the ice rheology.

avg_temps::F: Intermediate buffer for computing average temperatures.

avg_gradients::F: Intermediate buffer for computing average gradients.

In order to create Glacier types with information of a given glacier for a simulation, one can initialize a list of glaciers based on RGI (Randolph Glacier Inventory) IDs. Before running this, make sure to have used Gungnir to download all the necessary data for those glaciers, or double check that these glaciers are already available on the ODINN server.

Sleipnir.initialize_glaciers — Function

initialize_glaciers(
    rgi_ids::Vector{String},
    params::Parameters;
    velocityDatacubes::Union{Dict{String, String}, Dict{String, RasterStack}}=Dict(),
)

Initialize glaciers based on provided RGI IDs and parameters.

Arguments

rgi_ids::Vector{String}: A vector of RGI IDs representing the glaciers to be initialized.

params::Parameters: A Parameters object containing simulation parameters.

test::Bool: An optional boolean flag indicating whether to run in test mode. Default is false.

velocityDatacubes::Union{Dict{String, String}, Dict{String, RasterStack}}: A dictionary that provides for each RGI ID either the path to the datacube or the RasterStack with velocity data.

Returns

glaciers::Vector{Glacier2D}: A vector of initialized Glacier2D objects.

Description

This function performs the following steps:

Generates a file for missing glaciers if it does not already exist.

Filters out missing glaciers from the provided RGI IDs.

Generates raw climate data for the glaciers if necessary.

Initializes the glaciers using the provided RGI IDs and parameters.

If use_glathida_data is enabled in the simulation parameters, assigns GlaThiDa data to the glaciers.

Errors

Throws an error if none of the provided RGI IDs have GlaThiDa data.

Warnings

Issues a warning if not all glaciers have GlaThiDa data available.

Example

# We declare a list of glaciers to be initialized with their RGI IDs
rgi_ids = ["RGI60-11.03638", "RGI60-11.01450", "RGI60-11.02346", "RGI60-08.00203"]
# We initialize those glaciers based on the RGI IDs and the parameters we previously specified
glaciers = initialize_glaciers(rgi_ids, params)

Models

There are 3 main types of models in ODINN.jl, iceflow models, mass balance models and machine learning models. These three families are determined by abstract types, with specific types being declared as subtypes of these abstract types to ensure compatibility through the ODINN ecosystem.

The three main types of models are gathered in a type Model in the following way:

ODINN.Model — Function

Model(; iceflow::Union{IFM, Vector{IFM}, Nothing}, mass_balance::Union{MBM, Vector{MBM}, Nothing}, machine_learning::Union{MLM, Nothing}) where {IFM <: IceflowModel, MBM <: MBmodel, MLM <: MLmodel}

Creates a new model instance using the provided iceflow, mass balance, and machine learning components.

Arguments

iceflow::Union{IFM, Vector{IFM}, Nothing}: The iceflow model(s) to be used. Can be a single model, a vector of models, or nothing.

mass_balance::Union{MBM, Vector{MBM}, Nothing}: The mass balance model(s) to be used. Can be a single model, a vector of models, or nothing.

machine_learning::Union{MLM, Nothing}: The machine learning model to be used. Can be a single model or nothing.

Returns

model: A new instance of Sleipnir.Model initialized with the provided components.

ODINN.Model — Method

Model(; iceflow::Union{IFM, Vector{IFM}, Nothing}, mass_balance::Union{MBM, Vector{MBM}, Nothing}, machine_learning::Union{MLM, Nothing}) where {IFM <: IceflowModel, MBM <: MBmodel, MLM <: MLmodel}

Creates a new model instance using the provided iceflow, mass balance, and machine learning components.

Arguments

iceflow::Union{IFM, Vector{IFM}, Nothing}: The iceflow model(s) to be used. Can be a single model, a vector of models, or nothing.

mass_balance::Union{MBM, Vector{MBM}, Nothing}: The mass balance model(s) to be used. Can be a single model, a vector of models, or nothing.

machine_learning::Union{MLM, Nothing}: The machine learning model to be used. Can be a single model or nothing.

Returns

model: A new instance of Sleipnir.Model initialized with the provided components.

Ice flow models

Ice flow models are used to solve the PDEs describing the gravitational flow of glaciers. All ice flow models must be a subtype of abstract type IceflowModel. Ice flow models are managed by Huginn.jl.

The main type of ice flow model used in ODINN.jl right now is a 2D Shallow Ice Approximation (SIA) model (Hutter, 1983). This is declared in the following way:

Huginn.SIA2Dmodel — Type

mutable struct SIA2Dmodel{R <: Real, I <: Integer} <: SIAmodel

A mutable struct representing a 2D Shallow Ice Approximation (SIA) model.

Fields

A::Union{Ref{R}, Nothing}: Flow rate factor.

n::Union{Ref{R}, Nothing}: Flow law exponent.

C::Union{Ref{R}, Matrix{R}, Nothing}: Sliding coefficient.

H₀::Matrix{R}: Initial ice thickness.

H::Union{Matrix{R}, Nothing}: Ice thickness.

H̄::Union{Matrix{R}, Nothing}: Averaged ice thickness.

S::Matrix{R}: Surface elevation.

dSdx::Union{Matrix{R}, Nothing}: Surface slope in the x-direction.

dSdy::Union{Matrix{R}, Nothing}: Surface slope in the y-direction.

D::Union{Matrix{R}, Nothing}: Diffusivity.

D_is_provided::Union{Bool, Nothing}: Specifies if D is provided by user or computed.

Dx::Union{Matrix{R}, Nothing}: Diffusivity in the x-direction.

Dy::Union{Matrix{R}, Nothing}: Diffusivity in the y-direction.

dSdx_edges::Union{Matrix{R}, Nothing}: Surface slope at edges in the x-direction.

dSdy_edges::Union{Matrix{R}, Nothing}: Surface slope at edges in the y-direction.

∇S::Union{Matrix{R}, Nothing}: Gradient of the surface elevation.

∇Sy::Union{Matrix{R}, Nothing}: Gradient of the surface elevation in the y-direction.

∇Sx::Union{Matrix{R}, Nothing}: Gradient of the surface elevation in the x-direction.

Fx::Union{Matrix{R}, Nothing}: Flux in the x-direction.

Fy::Union{Matrix{R}, Nothing}: Flux in the y-direction.

Fxx::Union{Matrix{R}, Nothing}: Second derivative of flux in the x-direction.

Fyy::Union{Matrix{R}, Nothing}: Second derivative of flux in the y-direction.

V::Union{Matrix{R}, Nothing}: Velocity.

Vx::Union{Matrix{R}, Nothing}: Velocity in the x-direction.

Vy::Union{Matrix{R}, Nothing}: Velocity in the y-direction.

Γ::Union{Ref{R}, Nothing}: Basal shear stress.

MB::Union{Matrix{R}, Nothing}: Mass balance.

MB_mask::Union{AbstractArray{Bool}, Nothing}: Mask for mass balance.

MB_total::Union{Matrix{R}, Nothing}: Total mass balance.

glacier_idx::Union{Ref{I}, Nothing}: Index of the glacier.

Huginn.SIA2Dmodel — Method

SIA2Dmodel(
    params::Sleipnir.Parameters;
    A::Union{R, Nothing} = nothing,
    n::Union{R, Nothing} = nothing,
    C::Union{R, Matrix{R}, Nothing} = nothing,
    H₀::Matrix{R} = Matrix{Sleipnir.Float}([;;]),
    H::Union{Matrix{R}, Nothing} = nothing,
    H̄::Union{Matrix{R}, Nothing} = nothing,
    S::Matrix{R} = Matrix{Sleipnir.Float}([;;]),
    dSdx::Union{Matrix{R}, Nothing} = nothing,
    dSdy::Union{Matrix{R}, Nothing} = nothing,
    D::Union{Matrix{R}, Nothing} = nothing,
    D_is_provided::Union{Bool, Nothing} = nothing,
    Dx::Union{Matrix{R}, Nothing} = nothing,
    Dy::Union{Matrix{R}, Nothing} = nothing,
    dSdx_edges::Union{Matrix{R}, Nothing} = nothing,
    dSdy_edges::Union{Matrix{R}, Nothing} = nothing,
    ∇S::Union{Matrix{R}, Nothing} = nothing,
    ∇Sy::Union{Matrix{R}, Nothing} = nothing,
    ∇Sx::Union{Matrix{R}, Nothing} = nothing,
    Fx::Union{Matrix{R}, Nothing} = nothing,
    Fy::Union{Matrix{R}, Nothing} = nothing,
    Fxx::Union{Matrix{R}, Nothing} = nothing,
    Fyy::Union{Matrix{R}, Nothing} = nothing,
    V::Union{Matrix{R}, Nothing} = nothing,
    Vx::Union{Matrix{R}, Nothing} = nothing,
    Vy::Union{Matrix{R}, Nothing} = nothing,
    Γ::Union{R, Nothing} = nothing,
    MB::Union{Matrix{R}, Nothing} = nothing,
    MB_mask::Union{BitMatrix, Nothing} = nothing,
    MB_total::Union{Matrix{R}, Nothing} = nothing,
    glacier_idx::Union{I, Nothing} = nothing
) where {I <: Integer, R <: Real}

Constructs a new SIA2Dmodel object with the given parameters.

Arguments

params::Sleipnir.Parameters: Simulation parameters.

A::Union{R, Nothing}: Flow law parameter (default: nothing).

n::Union{R, Nothing}: Flow law exponent (default: nothing).

C::Union{R, Matrix{R}, Nothing}: Basal sliding parameter (default: nothing).

H₀::Matrix{R}: Initial ice thickness (default: empty matrix).

H::Union{Matrix{R}, Nothing}: Ice thickness (default: nothing).

H̄::Union{Matrix{R}, Nothing}: Averaged ice thickness (default: nothing).

S::Matrix{R}: Surface elevation (default: empty matrix).

dSdx::Union{Matrix{R}, Nothing}: Surface slope in x-direction (default: nothing).

dSdy::Union{Matrix{R}, Nothing}: Surface slope in y-direction (default: nothing).

D::Union{Matrix{R}, Nothing}: Diffusivity (default: nothing).

D_is_provided::Union{Bool, Nothing}: Specifies if D is provided by user or computed (default: false).

Dx::Union{Matrix{R}, Nothing}: Diffusivity in x-direction (default: nothing).

Dy::Union{Matrix{R}, Nothing}: Diffusivity in y-direction (default: nothing).

dSdx_edges::Union{Matrix{R}, Nothing}: Surface slope at edges in x-direction (default: nothing).

dSdy_edges::Union{Matrix{R}, Nothing}: Surface slope at edges in y-direction (default: nothing).

∇S::Union{Matrix{R}, Nothing}: Gradient of surface elevation (default: nothing).

∇Sy::Union{Matrix{R}, Nothing}: Gradient of surface elevation in y-direction (default: nothing).

∇Sx::Union{Matrix{R}, Nothing}: Gradient of surface elevation in x-direction (default: nothing).

Fx::Union{Matrix{R}, Nothing}: Flux in x-direction (default: nothing).

Fy::Union{Matrix{R}, Nothing}: Flux in y-direction (default: nothing).

Fxx::Union{Matrix{R}, Nothing}: Second derivative of flux in x-direction (default: nothing).

Fyy::Union{Matrix{R}, Nothing}: Second derivative of flux in y-direction (default: nothing).

V::Union{Matrix{R}, Nothing}: Velocity (default: nothing).

Vx::Union{Matrix{R}, Nothing}: Velocity in x-direction (default: nothing).

Vy::Union{Matrix{R}, Nothing}: Velocity in y-direction (default: nothing).

Γ::Union{R, Nothing}: Auxiliary matrix (default: nothing).

MB::Union{Matrix{R}, Nothing}: Mass balance (default: nothing).

MB_mask::Union{BitMatrix, Nothing}: Mask for mass balance (default: nothing).

MB_total::Union{Matrix{R}, Nothing}: Total mass balance (default: nothing).

glacier_idx::Union{I, Nothing}: Index of the glacier (default: nothing).

Returns

SIA2Dmodel: A new SIA2Dmodel object.

When a simulation will be run in ODINN.jl using an ice flow model, its related equation will be solved using OrdinaryDiffEq.jl. The related equation to a SIA2Dmodel is declared in its related util functions. Generally, these equations need to exist both in-place (to reduce memory allocations and ensure maximum performance, see example below) or out-of-place (to be more AD-friendly).

Huginn.SIA2D! — Function

SIA2D!(dH::Matrix{R}, H::Matrix{R}, simulation::SIM, t::R) where {R <:Real, SIM <: Simulation}

Simulates the evolution of ice thickness in a 2D shallow ice approximation (SIA) model. Works in-place.

Arguments

dH::Matrix{R}: Matrix to store the rate of change of ice thickness.

H::Matrix{R}: Matrix representing the ice thickness.

simulation::SIM: Simulation object containing model parameters and state.

t::R: Current simulation time.

Details

This function updates the ice thickness H and computes the rate of change dH using the shallow ice approximation in 2D. It retrieves necessary parameters from the simulation object, enforces positive ice thickness values, updates glacier surface altimetry, computes surface gradients, flux components, and flux divergence.

Notes

The function operates on a staggered grid for computing gradients and fluxes.

Surface elevation differences are capped using upstream ice thickness to impose boundary conditions.

The function modifies the input matrices dH and H in-place.

Mass balance models

Mass balance models are used to simulate the simplified thermodynamics of the forcing of the atmosphere on glaciers. As per ice flow models, all specific mass balance models needs to be a subtype of the abstract type MBmodel. Mass balance models are managed by Muninn.jl. For now, we have simple temperature-index models, with either one or two degree-day factors (DDFs) (Hock, 2003):

Muninn.TImodel1 — Type

TImodel1{F <: AbstractFloat}

A structure representing a temperature index model with degree-day factor and accumulation factor.

Keyword arguments

DDF::F: Degree-day factor, which is a coefficient used to convert temperature into melt.

acc_factor::F: Accumulation factor, which is a coefficient used to adjust the accumulation of mass.

Type Parameters

F: A subtype of AbstractFloat representing the type of the factors.

Muninn.TImodel1 — Method

TImodel1(params::Sleipnir.Parameters; DDF::F = 7.0/1000.0, acc_factor::F = 1.0/1000.0) where {F <: AbstractFloat}

Create a temperature index model with one degree-day factor (DDF) with the given parameters.

Arguments

params::Sleipnir.Parameters: The simulation parameters.

DDF::F: Degree-day factor (default is 7.0/1000.0).

acc_factor::F: Accumulation factor (default is 1.0/1000.0).

Returns

TI1_model: An instance of TImodel1 with the specified parameters.

Surface mass balance models are run in DiscreteCallbacks from OrdinaryDiffEq.jl, which enable the safe execution during the solving of a PDE in specificly prescribed time steps determined in the stepsfield in Sleipnir.SimulationParameters.

We soon plan to add compatibility with neural networks coming from the MassBalanceMachine, which should become the de facto surface mass balance model in the ODINN.jl ecosystem.

Machine Learning models

Machine learning models are used in the context of Universal Differential Equations (UDEs, Rackauckas et al., 2020) to parametrize or learn specific parts of differential equations. Machine Learning models are manage by ODINN.jl. As per the other types of models, all machine learning models need to be a subtype of the abstract type MLmodel. The default solution here is to use a neural network:

ODINN.NeuralNetwork — Type

NeuralNetwork{
    ChainType <: Lux.Chain,
    ComponentVectorType <: ComponentVector,
    NamedTupleType <: NamedTuple
} <: MLmodel

Feed-forward neural network.

Fields

architecture::ChainType: Flux.Chain neural network architecture

θ::ComponentVectorType: Neural network parameters

st::NamedTupleType: Neural network status

ODINN.NeuralNetwork — Method

NN(params::Parameters;
    architecture::Union{Flux.Chain, Nothing} = nothing,
    θ::Union{Vector{AbstractFloat}, Nothing} = nothing)

Creates a new feed-forward neural network.

Keyword arguments

architecture: Flux.Chain neural network architecture (optional)

θ: Neural network parameters (optional)

Simulations

One can run different types of simulations in ODINN.jl. Any specific type of simulation must be a subtype of Simulation. All simulations share the same common interface designed around multiple dispatch. Basically, once a simulation type has been created, one can easily run by calling run!(simulation).

The main types of simulations are the following ones:

Prediction

A prediction, also known as a forward simulation, is just a forward simulation given a model configuration, based on parameters, glaciers and models. These are managed in Huginn.jl, since they do not involve any inverse methods nor parameter optimization.

Huginn.Prediction — Type

Prediction <: Simulation

A mutable struct that represents a prediction simulation.

Fields

model::Sleipnir.Model: The model used for the prediction.

glaciers::Vector{Sleipnir.AbstractGlacier}: A vector of glaciers involved in the prediction.

parameters::Sleipnir.Parameters: The parameters used for the prediction.

results::Vector{Results}: A vector of results obtained from the prediction.

Huginn.Prediction — Method

Prediction(model::Sleipnir.Model, glaciers::Vector{G}, parameters::Sleipnir.Parameters) where {G <: Sleipnir.AbstractGlacier}

Create a Prediction object using the given model, glaciers, and parameters.

Arguments

model::Sleipnir.Model: The model used for prediction.

glaciers::Vector{G}: A vector of glacier objects, where each glacier is a subtype of Sleipnir.AbstractGlacier.

parameters::Sleipnir.Parameters: The parameters used for the prediction.

Returns

Prediction: A Prediction object based on the input values.

Inversion

An inversion optimises a given set of model parameters, based on a given target and an optimizer. These are handled by ODINN.jl.

ODINN.Inversion — Type

Inversion <: Simulation

A mutable struct that represents an inversion simulation.

Fields

model::Sleipnir.Model: The model used for the inversion.

glaciers::Vector{Sleipnir.AbstractGlacier}: A vector of glaciers involved in the inversion.

parameters::Sleipnir.Parameters: The parameters used for the inversion.

inversion::Vector{InversionResults}: A vector of results from the inversion.

ODINN.Inversion — Method

Inversion(model::Sleipnir.Model, glaciers::Vector{G}, parameters::Sleipnir.Parameters) where {G <: Sleipnir.AbstractGlacier}

Create an Inversion object using the provided model, glaciers, and parameters.

Arguments

model::Sleipnir.Model: The model to be used for the inversion.

glaciers::Vector{G}: A vector of glaciers, where each glacier is a subtype of Sleipnir.AbstractGlacier.

parameters::Sleipnir.Parameters: The parameters to be used for the inversion.

Returns

inversion: An Inversion object initialized with the provided model, glaciers, and parameters.

Functional inversion

A functional inversion is the inversion of the parameters of a regressor (e.g. a neural network), which parametrize a function that modulates a parameter or set of parameters in a given mechanistic model (e.g. the SIA).

ODINN.FunctionalInversion — Type

mutable struct FunctionalInversion <: Simulation

An object representing a functional inversion simulation (i.e. the inversion of a function using some data-driven regressor).

Fields

model::Sleipnir.Model: The model used for the simulation.

glaciers::Vector{Sleipnir.AbstractGlacier}: A vector of glaciers involved in the simulation.

parameters::Sleipnir.Parameters: The parameters used for the simulation.

results::Vector{Results}: A vector to store the results of the simulation.

ODINN.FunctionalInversion — Method

function FunctionalInversion(
    model::Sleipnir.Model,
    glaciers::Vector{G},
    parameters::Sleipnir.Parameters
) where {G <: Sleipnir.AbstractGlacier}

Constructor for FunctionalInversion struct with glacier model information, glaciers, and parameters.

Arguments

model::Sleipnir.Model: The model used for the simulation.

glaciers::Vector{G}: A vector of glaciers involved in the simulation.

parameters::Sleipnir.Parameters: The parameters used for the simulation.

Returns

FunctionalInversion: A new instance of the FunctionalInversion struct.

Results and plotting

Every Simulation type has an associated Results object(s), one for each one of the glaciers in the simulation. This object, as its name indicates, stores all the results of the simulation, which can be used for data anlysis and plotting. These types are handled by Sleipnir.jl.

Sleipnir.Results — Type

mutable struct Results{F <: AbstractFloat, I <: Int}

A mutable struct to store the results of simulations.

Fields

rgi_id::String: Identifier for the RGI (Randolph Glacier Inventory).

H::Vector{Matrix{F}}: Vector of matrices representing glacier ice thickness H over time.

H_glathida::Matrix{F}: Optional matrix for Glathida ice thicknesses.

H_ref::Union{Nothing, Vector{Matrix{F}}}: Reference data for ice thickness.

S::Matrix{F}: Glacier surface altimetry.

B::Matrix{F}: Glacier bedrock.

V::Matrix{F}: Glacier ice surface velocities.

Vx::Matrix{F}: x-component of the glacier ice surface velocity V.

Vy::Matrix{F}: y-component of the glacier ice surface velocity V.

V_ref::Union{Nothing, Matrix{F}}: Reference data for glacier ice surface velocities V.

Vx_ref::Union{Nothing, Matrix{F}}: Reference data for the x-component of the glacier ice surface velocity Vx.

Vy_ref::Union{Nothing, Matrix{F}}: Reference data for the y-component of the glacier ice surface velocity Vy.

Δx::F: Grid spacing in the x-direction.

Δy::F: Grid spacing in the y-direction.

lon::Union{Nothing, F}: Optional longitude value.

lat::Union{Nothing, F}: Optional latitude value.

nx::I: Number of grid points in the x-direction.

ny::I: Number of grid points in the y-direction.

tspan::Vector{F}: Time span of the simulation.

θ::Union{Nothing, ComponentArray{F}}: Machine learning model parameters.

loss::Union{Nothing, Vector{F}} Vector with evolution of loss function.

Sleipnir.Results — Method

Results(glacier::G, ifm::IF;
    rgi_id::String = glacier.rgi_id,
    H::Union{Nothing, Vector{Matrix{F}}} = nothing,
    H_glathida::Matrix{F} = glacier.H_glathida,
    H_ref::Union{Nothing, Vector{Matrix{F}}} = nothing,
    S::Union{Nothing, Matrix{F}} = nothing,
    B::Union{Nothing, Matrix{F}} = nothing,
    V::Union{Nothing, Vector{Matrix{F}}} = nothing,
    Vx::Union{Nothing, Vector{Matrix{F}}} = nothing,
    Vy::Union{Nothing, Vector{Matrix{F}}} = nothing,
    V_ref::Union{Nothing, Vector{Matrix{F}}} = nothing,
    Vx_ref::Union{Nothing, Vector{Matrix{F}}} = nothing,
    Vy_ref::Union{Nothing, Vector{Matrix{F}}} = nothing,
    Δx::F = glacier.Δx,
    Δy::F = glacier.Δy,
    lon::Union{Nothing, F} = glacier.cenlon,
    lat::Union{Nothing, F} = glacier.cenlat,
    nx::I = glacier.nx,
    ny::I = glacier.ny,
    t::Union{Vector{F}, Nothing} = nothing,
    tspan::Union{Tuple{F, F}, Nothing} = nothing,
    θ::Union{Nothing,ComponentArray{F}} = nothing,
    loss::Union{Nothing,Vector{F}} = nothing
) where {G <: AbstractGlacier, F <: AbstractFloat, IF <: AbstractModel, I <: Int}

Construct a Results object for a glacier simulation.

Arguments

glacier::G: The glacier object, subtype of AbstractGlacier.

ifm::IF: The model object, subtype of AbstractModel.

rgi_id::String: The RGI identifier for the glacier. Defaults to glacier.rgi_id.

H::Union{Nothing, Vector{Matrix{F}}}: Ice thickness matrices. Defaults to nothing.

H_glathida::Matrix{F}: Ice thickness from GlaThiDa. Defaults to glacier.H_glathida.

H_ref::Union{Nothing, Vector{Matrix{F}}}: Reference ice thickness. Defaults to nothing.

S::Union{Nothing, Matrix{F}}: Surface elevation matrix. Defaults to a zero matrix of the same size as ifm.S.

B::Union{Nothing, Matrix{F}}: Bed elevation matrix. Defaults to a zero matrix of the same size as glacier.B.

V::Union{Nothing, Vector{Matrix{F}}}: Velocity magnitude matrix. Defaults to nothing.

Vx::Union{Nothing, Vector{Matrix{F}}}: Velocity in the x-direction matrix. Defaults to nothing.

Vy::Union{Nothing, Vector{Matrix{F}}}: Velocity in the y-direction matrix. Defaults to nothing.

V_ref::Union{Nothing, Vector{Matrix{F}}}: Reference velocity magnitude matrix. Defaults to nothing.

Vx_ref::Union{Nothing, Vector{Matrix{F}}}: Reference velocity in the x-direction matrix. Defaults to nothing.

Vy_ref::Union{Nothing, Vector{Matrix{F}}}: Reference velocity in the y-direction matrix. Defaults to nothing.

Δx::F: Grid spacing in the x-direction. Defaults to glacier.Δx.

Δy::F: Grid spacing in the y-direction. Defaults to glacier.Δy.

lon::Union{Nothing, F}: Longitude of the glacier center. Defaults to glacier.cenlon.

lat::Union{Nothing, F}: Latitude of the glacier center. Defaults to glacier.cenlat.

nx::I: Number of grid points in the x-direction. Defaults to glacier.nx.

ny::I: Number of grid points in the y-direction. Defaults to glacier.ny.

tspan::Tuple(F, F): Timespan of the simulation.

θ::Union{Nothing, ComponentArray{F}}: Model parameters. Defaults to nothing.

loss::Union{Nothing, Vector{F}}: Loss values. Defaults to nothing.

Returns

results::Results: A Results object containing the simulation results.

One of the main things one can do with a Results object, is plotting them. The main function to do so is the following one:

Sleipnir.plot_glacier — Function

plot_glacier(results::Results, plot_type::String, variables::Vector{Symbol}; kwargs...) -> Figure

Generate various types of plots for glacier data.

Arguments

results::Results: The results object containing the data to be plotted.

plot_type::String: Type of plot to generate. Options are:

"heatmaps": Heatmaps for glacier variables like :H, :H₀, :S, :B, :V, :Vx, :Vy, :V_ref.

"evolution difference": Temporal difference metrics (between start and end) for a variable, with optional metrics like "hist" (histogram) and "difference".

"evolution statistics": Temporal statistical metrics for a variable, with optional metrics like "average", "median", "min", "max", and "std".

"integrated volume": Temporal evolution of the integrated ice volume for a variable.

"bias": Scatter plot to visualize the bias between two variables.

variables::Vector{Symbol}: Variables to be plotted, e.g., :H.

Optional Keyword Arguments

tspan: A tuple representing the start and end time for the simulation.

metrics: Metrics to visualize, e.g., ["average"] for statistics, ["difference"] for difference.

scale_text_size::Union{Nothing,Float64}: Optional argument to scale the text size for heatmaps.

threshold::Vector{F}: Threshold values for filtering data in bias plots.

figsize::Tuple{Int64, Int64}: Size of the figure.

Returns

A Figure object containing the desired visualization.

Notes

Ensure the variables and kwargs match the requirements of the specified plot_type.

The function routes requests to specific plotting functions based on plot_type.

Another option is to generate a video of the evolution of the glacier's ice thickness during the simulation:

Sleipnir.plot_glacier_vid — Function

plot_glacier_vid(
    plot_type::String,
    H::Vector{Matrix{Float64}},
    glacier::Glacier2D,
    simuparams::SimulationParameters,
    pathVideo::String;
    framerate::Int=24,
    baseTitle::String=""
)

Generate various types of videos for glacier data. For now only the evolution of the glacier ice thickness is supported. More types of visualizations will be added in the future.

Arguments

plot_type: Type of plot to generate. Options are:

"thickness": Heatmap of the glacier thickness.

H: A vector of matrices containing the ice thickness over time. This should be replaced by a Results instance in the future once Results no longer depends on an iceflow model.

glacier: A glacier instance.

simuparams: The simulation parameters.

pathVideo: Path of the mp4 file to generate.

Optional Keyword Arguments

framerate: The framerate to use for the video generation.

baseTitle: The prefix to use in the title of the frames. In each frame it is concatenated with the value of the year in the form " (t=XXXX)".