Simulations

One can run different types of simulations in ODINN.jl. Any specific type of simulation must be a subtype of Simulation. All simulations share the same common interface designed around multiple dispatch. Basically, once a simulation type has been created, one can easily run by calling run!(simulation).

The main types of simulations are the following ones:

Prediction

A prediction, also known as a forward simulation, is just a forward simulation given a model configuration, based on parameters, glaciers and models. These are managed in Huginn.jl, since they do not involve any inverse methods nor parameter optimization.

Huginn.Prediction — Type

Prediction{CACHE} <: Simulation

A mutable struct that represents a prediction simulation.

Fields

model::Sleipnir.Model: The model used for the prediction.

glaciers::Vector{Sleipnir.AbstractGlacier}: A vector of glaciers involved in the prediction.

parameters::Sleipnir.Parameters: The parameters used for the prediction.

results::Vector{Results}: A vector of results obtained from the prediction.

Huginn.Prediction — Method

Prediction(model::Sleipnir.Model, glaciers::Vector{G}, parameters::Sleipnir.Parameters) where {G <: Sleipnir.AbstractGlacier}

Create a Prediction object using the given model, glaciers, and parameters.

Arguments

model::Sleipnir.Model: The model used for prediction.

glaciers::Vector{G}: A vector of glacier objects, where each glacier is a subtype of Sleipnir.AbstractGlacier.

parameters::Sleipnir.Parameters: The parameters used for the prediction.

Returns

Prediction: A Prediction object based on the input values.

Inversion

An Inversion is the inversion of the parameters involved in the PDE, or the parameters of a regressor (e.g. a neural network), which parametrize a function that modulates a parameter or set of parameters in a given mechanistic model (e.g. the SIA).

ODINN.Inversion — Type

mutable struct Inversion{MODEL, CACHE, GLACIER, RES} <: Simulation

An object representing an inversion simulation. It can involve at the same time a classical inversion and a functional inversion (i.e. the inversion of a function using some data-driven regressor).

Fields

model::Sleipnir.Model: The model used for the simulation.

glaciers::Vector{Sleipnir.AbstractGlacier}: A vector of glaciers involved in the simulation.

parameters::Sleipnir.Parameters: The parameters used for the simulation.

results::ODINN.Results: A ODINN.Results instance to store the results of the inversion and of the forward simulations.

ODINN.Inversion — Method

function Inversion(
    model::M,
    glaciers::Vector{G},
    parameters::P
) where {G <: Sleipnir.AbstractGlacier, M <: Sleipnir.Model, P <: Sleipnir.Parameters}

Constructor for Inversion struct with glacier model information, glaciers, and parameters.

Arguments

model::Sleipnir.Model: The model used for the simulation.

glaciers::Vector{G}: A vector of glaciers involved in the simulation.

parameters::Sleipnir.Parameters: The parameters used for the simulation.

Returns

Inversion: A new instance of the Inversion struct.